### 2022-10-06 #4820 @editorialbot
[PRE REVIEW]: icepyx: querying, obtaining, analyzing, and manipulating ICESat-2 datasets
**Submitting author:** @JessicaS11 (<a href="http://orcid.org/0000-0002-3421-4459">Jessica Scheick</a>)
**Repository:** https://github.com/icesat2py/icepyx
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.6.3
