### 2022-10-06 6bc8590
add joss bot recommended doi
