+++ doc/source/tracking/icepyx_pubs.bib
@@ -35,11 +35,16 @@ doi          = {10.5281/zenodo.6462479},
 url          = {https://doi.org/10.5281/zenodo.6462479}
 }
 
-@misc{js2021agu,
-author = {Scheick, J. and Bisson, K. and Li, T. and Leong, W.J. and Arendt, A.},
+@article{js2021agu,
+author = {Scheick, Jessica and Bisson, Kelsey and Li, Tian and Leong, Wei Ji and Arendt, Anthony},
 note = {Abstract and poster (presented by Wei Ji Leong). American Geophysical Union Fall Meeting, New Orleans, LA, USA. 13-17 December 2021.},
 title = {Collaborative Computational Resource Development around {ICESat-2} Data: the {icepyx} Community and Library},
-year = {2021},
+journal = {Earth and Space Science Open Archive},
+pages = {9},
+year = {2022}, 
+% for AGU Fall Meeting 2021
+DOI = {10.1002/essoar.10511316.1},
+url = {https://doi.org/10.1002/essoar.10511316.1},
 comment = {abstract, poster},
 }
 
