### 2022-10-06 2b1427a
debug spatial tests
