+++ icepyx/tests/test_spatial.py
@@ -131,9 +131,18 @@ def test_list_pairs_polygon():
     poly_list_pair = spat.Spatial(
         [[-55, 68], [-55, 71], [-48, 71], [-48, 68], [-55, 68]]
     )
-    expected_poly_list_pair = Polygon(
-        [[-55, 68], [-55, 71], [-48, 71], [-48, 68], [-55, 68]]
-    )
+    expected_poly_list_pair = [
+        -55.0,
+        68.0,
+        -55.0,
+        71.0,
+        -48.0,
+        71.0,
+        -48.0,
+        68.0,
+        -55.0,
+        68.0,
+    ]
 
     assert poly_list_pair._ext_type == "polygon"
     assert poly_list_pair._geom_file is None
@@ -144,9 +153,18 @@ def test_tuple_latlon_pairs():
     poly_tuple_pair = spat.Spatial(
         [(-55, 68), (-55, 71), (-48, 71), (-48, 68), (-55, 68)]
     )
-    expected_poly_tuple_pair = Polygon(
-        [[-55, 68], [-55, 71], [-48, 71], [-48, 68], [-55, 68]]
-    )
+    expected_poly_tuple_pair = [
+        -55.0,
+        68.0,
+        -55.0,
+        71.0,
+        -48.0,
+        71.0,
+        -48.0,
+        68.0,
+        -55.0,
+        68.0,
+    ]
 
     assert poly_tuple_pair._ext_type == "polygon"
     assert poly_tuple_pair._geom_file is None
@@ -155,11 +173,19 @@ def test_tuple_latlon_pairs():
 
 def test_intlist_latlon_coords():
     poly_list = spat.Spatial([-55, 68, -55, 71, -48, 71, -48, 68, -55, 68])
-    expected_poly_list = Polygon(
-        [[-55, 68], [-55, 71], [-48, 71], [-48, 68], [-55, 68]]
-    )
-    print(poly_list.extent)
-    print(expected_poly_list)
+    expected_poly_list = [
+        -55.0,
+        68.0,
+        -55.0,
+        71.0,
+        -48.0,
+        71.0,
+        -48.0,
+        68.0,
+        -55.0,
+        68.0,
+    ]
+
     assert poly_list._ext_type == "polygon"
     assert poly_list._geom_file is None
     assert poly_list.extent == expected_poly_list
@@ -169,9 +195,18 @@ def test_floatlist_latlon_coords():
     poly_float_list = spat.Spatial(
         [-55.0, 68.7, -55.0, 71, -48, 71, -48, 68.7, -55.0, 68.7]
     )
-    expected_poly_float_list = Polygon(
-        [[-55.0, 68.7], [-55.0, 71], [-48, 71], [-48, 68.7], [-55.0, 68.7]]
-    )
+    expected_poly_float_list = [
+        -55.0,
+        68.7,
+        -55.0,
+        71.0,
+        -48.0,
+        71.0,
+        -48.0,
+        68.7,
+        -55.0,
+        68.7,
+    ]
 
     assert poly_float_list._ext_type == "polygon"
     assert poly_float_list._geom_file is None
@@ -185,9 +220,18 @@ def test_numpy_list_pairs_polygon():
     poly_list_pair = spat.Spatial(
         np.array([[-55, 68], [-55, 71], [-48, 71], [-48, 68], [-55, 68]])
     )
-    expected_poly_list_pair = Polygon(
-        [[-55, 68], [-55, 71], [-48, 71], [-48, 68], [-55, 68]]
-    )
+    expected_poly_list_pair = [
+        -55.0,
+        68.0,
+        -55.0,
+        71.0,
+        -48.0,
+        71.0,
+        -48.0,
+        68.0,
+        -55.0,
+        68.0,
[…]
