+++ doc/source/tracking/paper.md
@@ -40,7 +40,7 @@ authors:
     orcid: 0000-0002-1577-4004
     affiliation: 7
   - name: Zheng Liu
-    orcid: 0000-0003-4132-813
+    orcid: 0000-0003-4132-8136
     affiliation: 4
   - name: Romina Piunno
     affiliation: 8
