### 2022-10-07 8a1870e
clean up some dev notes
