### 2022-10-07 c70c9cf
keep working on setting spatial extent for file inputs
