+++ icepyx/tests/test_APIformatting.py
@@ -30,974 +30,6 @@ def test_temporal_fmt():
     assert obs == exp
 
 
-########## _fmt_spatial ##########
-def test_bounding_box_fmt():
-    obs = apifmt._fmt_spatial("bounding_box", [-55, 68, -48, 71])
-    exp = {"bounding_box": "-55,68,-48,71"}
-    assert obs == exp
-
-
-def test_bbox_fmt():
-    obs = apifmt._fmt_spatial("bbox", [-55, 68, -48, 71])
-    exp = {"bbox": "-55,68,-48,71"}
-    assert obs == exp
-
-
-def test_polygon_fmt():
-    poly = Polygon(
-        [
-            [-86.622742, -74.908126, 0.0],
-            [-86.602149, -74.998483, 0.0],
-            [-86.671945, -74.999545, 0.0],
-            [-86.667881, -75.01762, 0.0],
-            [-86.737771, -75.018662, 0.0],
-            [-86.717729, -75.109052, 0.0],
-            [-86.788057, -75.110077, 0.0],
-            [-86.780144, -75.14624, 0.0],
-            [-86.850654, -75.147247, 0.0],
-            [-86.835058, -75.219586, 0.0],
-            [-86.905925, -75.220574, 0.0],
-            [-86.894389, -75.274839, 0.0],
-            [-86.965529, -75.27581, 0.0],
-            [-86.950368, -75.348177, 0.0],
-            [-87.021872, -75.349129, 0.0],
-            [-87.003154, -75.439609, 0.0],
-            [-87.075115, -75.440545, 0.0],
-            [-87.052886, -75.549149, 0.0],
-            [-86.98038, -75.548205, 0.0],
-            [-86.965004, -75.620616, 0.0],
-            [-87.037878, -75.621564, 0.0],
-            [-87.034102, -75.63967, 0.0],
-            [-86.961136, -75.63872, 0.0],
-            [-86.957257, -75.656825, 0.0],
-            [-86.884208, -75.655851, 0.0],
-            [-86.872234, -75.710165, 0.0],
-            [-86.945563, -75.711143, 0.0],
-            [-86.925871, -75.801686, 0.0],
-            [-86.999681, -75.802647, 0.0],
-            [-86.988029, -75.856983, 0.0],
-            [-87.062135, -75.857925, 0.0],
-            [-87.058326, -75.87604, 0.0],
-            [-87.132537, -75.876959, 0.0],
-            [-87.128815, -75.895075, 0.0],
-            [-87.203132, -75.895972, 0.0],
-            [-87.199496, -75.914091, 0.0],
-            [-87.273919, -75.914965, 0.0],
-            [-87.27037, -75.933086, 0.0],
-            [-87.344899, -75.933938, 0.0],
-            [-87.341438, -75.95206, 0.0],
-            [-87.490715, -75.953695, 0.0],
-            [-87.487439, -75.97182, 0.0],
-            [-87.562188, -75.972604, 0.0],
-            [-87.559001, -75.99073, 0.0],
-            [-87.633856, -75.991491, 0.0],
-            [-87.627653, -76.027748, 0.0],
-            [-87.777778, -76.029203, 0.0],
-            [-87.774861, -76.047334, 0.0],
-            [-87.925213, -76.048696, 0.0],
-            [-87.922485, -76.06683, 0.0],
-            [-87.997771, -76.067476, 0.0],
-            [-87.995135, -76.085611, 0.0],
-            [-88.070527, -76.086234, 0.0],
-            [-88.067983, -76.104371, 0.0],
-            [-88.143481, -76.104971, 0.0],
-            [-88.14103, -76.123109, 0.0],
-            [-88.216633, -76.123686, 0.0],
-            [-88.214276, -76.141826, 0.0],
[…]
