+++ icepyx/core/granules.py
@@ -171,7 +171,7 @@ class Granules:
     # ----------------------------------------------------------------------
     # Methods
 
-    def get_avail(self, CMRparams, reqparams):
+    def get_avail(self, CMRparams, reqparams, cloud=False):
         """
         Get a list of available granules for the query object's parameters.
         Generates the `avail` attribute of the granules object.
@@ -183,6 +183,8 @@ class Granules:
         reqparams : dictionary
             Dictionary of properly formatted parameters required for searching, ordering,
             or downloading from NSIDC.
+        cloud : boolean, default False
+            Wether or not you want data available in the cloud (versus on premises).
 
         Notes
         -----
@@ -207,8 +209,13 @@ class Granules:
         headers = {"Accept": "application/json", "Client-Id": "icepyx"}
         # note we should also check for errors whenever we ping NSIDC-API - make a function to check for errors
 
+        if cloud == True:
+            prov_flag = "NSIDC_CPRD"
+        else:
+            prov_flag = "NSIDC_ECS"
+
         params = apifmt.combine_params(
-            CMRparams, {k: reqparams[k] for k in ["page_size"]}
+            CMRparams, {k: reqparams[k] for k in ["page_size"]}, {"provider": prov_flag}
         )
 
         cmr_search_after = None
