### 2022-10-17 b5ee2a7
Update icepyx/core/granules.py
Co-authored-by: Tyler Sutterley <tsutterl@uw.edu>
