### 2022-10-18 72210b4
fix rest of geospatial test issues
