+++ icepyx/core/spatial.py
@@ -49,7 +49,7 @@ def geodataframe(extent_type, spatial_extent, file=False, xdateline=None):
     >>> reg_a = ipx.Query('ATL06',[-55, 68, -48, 71],['2019-02-20','2019-02-28'])
     >>> gdf = geodataframe(reg_a.spatial.extent_type, reg_a.spatial.extent)
     >>> gdf.geometry
-    0   POLYGON ((-55.00000 68.00000, -55.00000 71.000...
+    0    POLYGON ((-48.00000 68.00000, -48.00000 71.000...
     Name: geometry, dtype: geometry
     """
 
@@ -661,7 +661,7 @@ class Spatial:
 
         EGI spatial inputs must be formatted a specific way.
         This method formats the given spatial extent to be a valid submission.
-        For large/complex polygons, this includes simplifying the polygon.
+        DevGoal: For large/complex polygons, this includes simplifying the polygon.
         Coordinates will be properly ordered, and the required string formatting applied.
 
         Returns
