+++ icepyx/tests/test_APIformatting.py
@@ -138,17 +138,17 @@ def test_CMRparams_no_other_inputs():
 
     CMRparams.build_params(
         product="ATL06",
-        version="003",
+        version="005",
         start=dt.datetime(2019, 2, 20, 0, 0),
         end=dt.datetime(2019, 2, 24, 23, 59, 59),
         extent_type="bounding_box",
-        spatial_extent=[-55, 68, -48, 71],
+        spatial_extent="-55.0,68.0,-48.0,71.0",
     )
     obs_fmted_params = CMRparams.fmted_keys
     exp_fmted_params = {
         "short_name": "ATL06",
-        "version": "003",
+        "version": "005",
         "temporal": "2019-02-20T00:00:00Z,2019-02-24T23:59:59Z",
-        "bounding_box": "-55,68,-48,71",
+        "bounding_box": "-55.0,68.0,-48.0,71.0",
     }
     assert obs_fmted_params == exp_fmted_params
