### 2022-10-18 f4128ec
Remove upper pin for python_requires
