### 2022-10-19 9edc83e
code and notebook updates for cloud access
