+++ icepyx/core/granules.py
@@ -35,7 +35,7 @@ def info(grans):
 
 # DevNote: currently this fn is not tested
 # DevNote: could add flag to separate ascending and descending orbits based on ATL03 granule region
-def gran_IDs(grans, ids=True, cycles=False, tracks=False, dates=False, s3urls=False):
+def gran_IDs(grans, ids=True, cycles=False, tracks=False, dates=False, cloud=False):
     """
     Returns a list of granule information for each granule dictionary in the input list of granule dictionaries.
     Granule info may be from a list of those available from NSIDC (for ordering/download)
@@ -53,7 +53,7 @@ def gran_IDs(grans, ids=True, cycles=False, tracks=False, dates=False, s3urls=Fa
         Return a list of the available Reference Ground Tracks (RGTs) for the granule dictionary
     dates : boolean, default False
         Return a list of the available dates for the granule dictionary.
-    s3urls : boolean, default False
+    cloud : boolean, default False
         Return a a list of AWS s3 urls for the available granules in the granule dictionary.
         Note: currently, NSIDC does not provide metadata on which granules are available on s3.
         Thus, all of the urls may not be valid and may return FileNotFoundErrors.
@@ -76,8 +76,6 @@ def gran_IDs(grans, ids=True, cycles=False, tracks=False, dates=False, s3urls=Fa
 
         if int(gran["producer_granule_id"][3:5]) > 13:
             continue
-            # ultimately use this to get the s3urls from the metadata
-            # gran_s3urls.append(gran["links"][])
 
         else:
             # PRD: ICESat-2 product
@@ -112,9 +110,10 @@ def gran_IDs(grans, ids=True, cycles=False, tracks=False, dates=False, s3urls=Fa
             gran_dates.append(
                 str(datetime.datetime(year=int(YY), month=int(MM), day=int(DD)).date())
             )
-            gran_s3urls.append(
-                f"s3://nsidc-cumulus-prod-protected/ATLAS/{PRD}/{RL}/{YY}/{MM}/{DD}/{producer_granule_id}"
-            )
+
+            for link in gran["links"]:
+                if link["href"].startswith("s3") and link["href"].endswith(".h5"):
+                    gran_s3urls.append(link["href"])
     # list of granule parameters
     gran_list = []
     # granule IDs
@@ -130,12 +129,7 @@ def gran_IDs(grans, ids=True, cycles=False, tracks=False, dates=False, s3urls=Fa
     if dates:
         gran_list.append(gran_dates)
     # AWS s3 url
-    if s3urls:
-        warnings.filterwarnings("always")
-        warnings.warn(
-            "You MUST be pre-authenticated by NSIDC as a beta tester to have cloud access to ICESat-2 data",
-            UserWarning,
-        )
+    if cloud:
         gran_list.append(gran_s3urls)
     # return the list of granule parameters
     return gran_list
