### 2022-10-19 870e944
Support Python 3.10 and 3.11 (#372)
* Support Python 3.10

Set upper pin of Python to be 3.10.

* Update PyPI classifiers and set development status from planning to beta

Ref: https://pypi.org/classifiers/

* Run readthedocs build on Python 3.10 and ubuntu 22.04

* Set upper Python version to 3.11

* Remove upper pin for python_requires

* GitHub action UML generation auto-update

Co-authored-by: GitHub Action <JessicaS11@users.noreply.github.com>
