### 2022-10-24 fcb2422
creates df with only profiles that contain all user specified params Need to dload additional params
