### 2022-10-24 a7b4a51
add sorting to pypi stats manipulations
