### 2022-10-25 #4820 @kthyng
Hi @JessicaS11 and thanks for your submission! I'll take this on as editor.

Could you please look through our list of [volunteer reviewers](https://bit.ly/joss-reviewers) and suggest people by github handle (without the "@") that look relevant? Also please feel free to suggest from outside this list.
