+++ icepyx/tests/test_is2ref.py
@@ -4,12 +4,6 @@ import warnings
 import icepyx.core.is2ref as is2ref
 
 ########## _validate_product ##########
-def test_lowercase_product():
-    lcds = "atl03"
-    obs = is2ref._validate_product(lcds)
-    expected = "ATL03"
-    assert obs == expected
-
 
 def test_num_product():
     dsnum = 6
@@ -23,6 +17,263 @@ def test_bad_product():
     ermsg = "Please enter a valid product"
     with pytest.raises(AssertionError, match=ermsg):
         is2ref._validate_product(wrngds)
+# test all possible product types
+def test_atl01_product():
+    lcds = "atl01"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL01"
+    assert obs == expected
+
+    ucds = "ATL01"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL01"
+    assert obs == expected
+
+
+def test_atl02product():
+    lcds = "atl02"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL02"
+    assert obs == expected
+
+    ucds = "ATL02"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL02"
+    assert obs == expected
+
+
+def test_atl03_product():
+    lcds = "atl03"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL03"
+    assert obs == expected
+
+    ucds = "ATL03"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL03"
+    assert obs == expected
+
+
+def test_atl04_product():
+    lcds = "atl04"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL04"
+    assert obs == expected
+
+    ucds = "ATL04"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL04"
+    assert obs == expected
+
+
+def test_atl06_product():
+    lcds = "atl06"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL06"
+    assert obs == expected
+
+    ucds = "ATL06"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL06"
+    assert obs == expected
+
+
+def test_atl07_product():
+    lcds = "atl07"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL07"
+    assert obs == expected
+
+    ucds = "ATL07"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL07"
+    assert obs == expected
+
+
+def test_atl07ql_product():
+    lcds = "atl07ql"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL07QL"
+    assert obs == expected
+
+    ucds = "ATL07QL"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL07QL"
+    assert obs == expected
+
+
+def test_atl08_product():
+    lcds = "atl08"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL08"
+    assert obs == expected
+
+    ucds = "ATL08"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL08"
+    assert obs == expected
+
+
+def test_atl09_product():
+    lcds = "atl09"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL09"
+    assert obs == expected
+
+    ucds = "ATL09"
+    obs = is2ref._validate_product(ucds)
+    expected = "ATL09"
+    assert obs == expected
+
+
+def test_atl09ql_product():
+    lcds = "atl09ql"
+    obs = is2ref._validate_product(lcds)
+    expected = "ATL09QL"
+    assert obs == expected
+
[…]
