+++ icepyx/tests/test_query.py
@@ -6,6 +6,8 @@ import warnings
 # 		Generic Query tests
 # ------------------------------------
 
+# seem to be adequately covered in docstrings;
+# may want to focus on testing specific queries
 
 # ------------------------------------
 # 		Icepyx-specific tests
