### 2022-10-26 9ddec8f
Apply suggestions from code review
