+++ icepyx/core/query.py
@@ -38,7 +38,7 @@ class GenQuery:
         Spatial extent of interest, provided as a bounding box, list of polygon coordinates, or
         geospatial polygon file.
         You can submit at most one bounding box or list of polygon coordinates.
-        Geospatial polygon files may contain multiple features (polygons).
+        Per NSIDC requirements, geospatial polygon files may only contain one feature (polygon).
         Bounding box coordinates should be provided in decimal degrees as
         [lower-left-longitude, lower-left-latitute, upper-right-longitude, upper-right-latitude].
         Polygon coordinates should be provided as coordinate pairs in decimal degrees as
