### 2022-10-26 1490c05
add earthdata as dependency
