### 2022-10-27 a135370
add some longitude range notes to docstring
