### 2022-10-28 3e38c5c
add cross-dateline handling and work on new spatial class (#331)
* add cross-dateline flag; check for dateline crossing and handling in geodataframe function
* move cmr and egi spatial formatting to spatial class to leverage class object info
* make use of private variables consistent
* clean up test failures due to properties/var naming
* add multi-feature handling in CMR
* add more edge cases and update accepted dateline check inputs

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
