+++ icepyx/core/spatial.py
@@ -2,7 +2,8 @@ import geopandas as gpd
 import numpy as np
 import os
 from pathlib import Path
-from shapely.geometry import Polygon
+from shapely.geometry import box, Polygon
+from shapely.geometry.polygon import orient
 import warnings
 
 import icepyx.core.APIformatting as apifmt
@@ -10,7 +11,7 @@ import icepyx.core.APIformatting as apifmt
 # DevGoal: need to update the spatial_extent docstring to describe coordinate order for input
 
 
-def geodataframe(extent_type, spatial_extent, file=False):
+def geodataframe(extent_type, spatial_extent, file=False, xdateline=None):
     """
     Return a geodataframe of the spatial extent
 
@@ -32,6 +33,13 @@ def geodataframe(extent_type, spatial_extent, file=False):
     file : boolean, default False
         Indication for whether the spatial_extent string is a filename or coordinate list
 
+    Returns
+    -------
+    gdf : GeoDataFrame
+        Returns a GeoPandas GeoDataFrame containing the spatial extent.
+        The GeoDataFrame will have only one entry unless a geospatial file
+        was submitted.
+
     See Also
     --------
     icepyx.Query
@@ -39,31 +47,34 @@ def geodataframe(extent_type, spatial_extent, file=False):
     Examples
     --------
     >>> reg_a = ipx.Query('ATL06',[-55, 68, -48, 71],['2019-02-20','2019-02-28'])
-    >>> gdf = geodataframe(reg_a.spatial.extent_type, reg_a.spatial.spatial_extent)
+    >>> gdf = geodataframe(reg_a.spatial.extent_type, reg_a.spatial.extent)
     >>> gdf.geometry
-    0   POLYGON ((-55.00000 68.00000, -55.00000 71.000...
+    0    POLYGON ((-48.00000 68.00000, -48.00000 71.000...
     Name: geometry, dtype: geometry
     """
 
+    if xdateline is not None:
+        xdateline = xdateline
+    elif file:
+        pass
+    else:
+        xdateline = check_dateline(extent_type, spatial_extent)
+    # print("this should cross the dateline:" + str(xdateline))
+
     if extent_type == "bounding_box":
-        boxx = [
-            spatial_extent[0],
-            spatial_extent[0],
-            spatial_extent[2],
-            spatial_extent[2],
-            spatial_extent[0],
-        ]
-        boxy = [
-            spatial_extent[1],
-            spatial_extent[3],
-            spatial_extent[3],
-            spatial_extent[1],
-            spatial_extent[1],
-        ]
-        # DevGoal: check to see that the box is actually correctly constructed;
-        # have not checked actual location of test coordinates
+        if xdateline == True:
+            cartesian_lons = [i if i > 0 else i + 360 for i in spatial_extent[0:-1:2]]
+            cartesian_spatial_extent = [
+                item
+                for pair in zip(cartesian_lons, spatial_extent[1::2])
+                for item in pair
+            ]
+            bbox = box(*cartesian_spatial_extent)
+        else:
+            bbox = box(*spatial_extent)
+
         # TODO: test case that ensures gdf is constructed as expected (correct coords, order, etc.)
-        gdf = gpd.GeoDataFrame(geometry=[Polygon(list(zip(boxx, boxy)))])
+        gdf = gpd.GeoDataFrame(geometry=[bbox], crs="epsg:4326")
 
     # DevGoal: Currently this if/else within this elif are not tested...
     # DevGoal: the crs setting and management needs to be improved
@@ -76,6 +87,16 @@ def geodataframe(extent_type, spatial_extent, file=False):
 
         # else, spatial_extent must be a list of floats (or list of tuples of floats)
         else:
+            if xdateline == True:
[…]
