### 2022-10-28 dfa21eb
update tracking + bib data (#383)
* add sorting to pypi stats manipulations
* update bib list
