### 2022-10-28 5f6c354
add deprecation to release log
