### 2022-10-28 db717b2
add try for key error when granids is used to parse incomplete metadata
