### 2022-10-28 5109fc5
update cloud tutorial to work with publicly accessible cloud data (#381)
* code and notebook updates for cloud access
* change `s3urls` keyword to `cloud`
