+++ icepyx/tests/test_behind_NSIDC_API_login.py
@@ -13,7 +13,7 @@ import warnings
 @pytest.fixture(scope="module")
 def reg():
     live_reg = ipx.Query(
-        "ATL06", [-55, 68, -48, 71], ["2019-02-22", "2019-02-28"], version="004"
+        "ATL06", [-55, 68, -48, 71], ["2019-02-22", "2019-02-28"], version="005"
     )
     yield live_reg
     del live_reg
@@ -39,8 +39,8 @@ import json
 
 
 def test_get_custom_options_output(session):
-    obs = is2ref._get_custom_options(session, "ATL06", "004")
-    with open("ATL06v04_options.json") as exp_json:
+    obs = is2ref._get_custom_options(session, "ATL06", "005")
+    with open("./icepyx/tests/ATL06v04_options.json") as exp_json:
         exp = json.load(exp_json)
         assert all(keys in obs.keys() for keys in exp.keys())
         assert all(obs[key] == exp[key] for key in exp.keys())
