+++ doc/source/user_guide/changelog/v0.6.4.rst
@@ -35,6 +35,7 @@ Maintenance
 - add provider flag to CMR request (needed with new cloud availability) (#380)
 - update traffic data and scripts (#363, #369, #383)
 - enable support for Python 3.10 and 3.11 (#372)
+- update json file with subsetting options to v005 of ATL06 (#387)
 
 
 Documentation
