### 2022-10-31 b111ced
fix filename in script
