+++ icepyx/tests/test_behind_NSIDC_API_login.py
@@ -40,7 +40,7 @@ import json
 
 def test_get_custom_options_output(session):
     obs = is2ref._get_custom_options(session, "ATL06", "005")
-    with open("./icepyx/tests/ATL06v04_options.json") as exp_json:
+    with open("./icepyx/tests/ATL06v05_options.json") as exp_json:
         exp = json.load(exp_json)
         assert all(keys in obs.keys() for keys in exp.keys())
         assert all(obs[key] == exp[key] for key in exp.keys())
