### 2022-11-04 #4820 @kthyng
Hi @dimejimudele and @vaasuDevans! Are you interested in reviewing this submission to JOSS? JOSS reviews are conducted in a github issue on this repo, and use a checklist. We ask you to open issues on the software repository to bring up specific comments/questions and stay organized. We like reviews to be done within 4–6 weeks but because this is an iterative process usually between the authors and reviewers, it's best to start earlier than that. Thank for your consideration!
