### 2022-11-07 #4912 @kthyng
👋🏼 @JessicaS11 @dimejimudele, @VaasuDevanS this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 2-4 weeks, or 4-6 weeks if needed. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
