### 2022-11-07 #4912 @weiji14
Just a quick note: new icepyx version 0.6.4 has been released on 31 Oct 2022, see https://github.com/icesat2py/icepyx/releases/tag/v0.6.4
