### 2022-11-07 88e5116
Remove comment on AGU Fall Meeting 2021 citation
Was causing issues with editorialbot at https://github.com/openjournals/joss-reviews/issues/4912#issuecomment-1305889442
