+++ doc/source/tracking/icepyx_pubs.bib
@@ -41,8 +41,7 @@ note = {Abstract and poster (presented by Wei Ji Leong). American Geophysical Un
 title = {Collaborative Computational Resource Development around {ICESat-2} Data: the {icepyx} Community and Library},
 journal = {Earth and Space Science Open Archive},
 pages = {9},
-year = {2022}, 
-% for AGU Fall Meeting 2021
+year = {2022},
 DOI = {10.1002/essoar.10511316.1},
 url = {https://doi.org/10.1002/essoar.10511316.1},
 comment = {abstract, poster},
@@ -160,18 +159,18 @@ abstract = {The Ice, Cloud, and land Elevation Satellite (ICESat) mission used l
 % #########  Research/Articles using (but not citing) icepyx ################
 
 
-@article{Sothe:2022, 
-title={Spatially Continuous Mapping of Forest Canopy Height in Canada by Combining GEDI and ICESat-2 with PALSAR and Sentinel}, 
-volume={14}, 
-ISSN={2072-4292}, 
-url={http://dx.doi.org/10.3390/rs14205158}, 
-DOI={10.3390/rs14205158}, 
-number={20}, 
-journal={Remote Sensing}, 
-publisher={MDPI AG}, 
-author={Sothe, Camile and Gonsamo, Alemu and Lourenço, Ricardo B. and Kurz, Werner A. and Snider, James}, 
-year={2022}, 
-month={Oct}, 
+@article{Sothe:2022,
+title={Spatially Continuous Mapping of Forest Canopy Height in Canada by Combining GEDI and ICESat-2 with PALSAR and Sentinel},
+volume={14},
+ISSN={2072-4292},
+url={http://dx.doi.org/10.3390/rs14205158},
+DOI={10.3390/rs14205158},
+number={20},
+journal={Remote Sensing},
+publisher={MDPI AG},
+author={Sothe, Camile and Gonsamo, Alemu and Lourenço, Ricardo B. and Kurz, Werner A. and Snider, James},
+year={2022},
+month={Oct},
 pages={5158},
 note={Attribution to icepyx from 15 Oct 2022 Twitter post, https://twitter.com/rblourenco/status/1581320878511382528}
 }
