### 2022-11-16 ba44f82
add profile (vs spot) handling for read-in
