### 2022-11-21 eb9c8ae
modified to populate prof df by querying individual profiles
