### 2022-11-22 1e20ed6
remove sc_orient from atl11 required vars list
