### 2022-11-22 a452ef9
introduce new function to determine if data product uses ground tracks, paths, or profiles and begin generalizing to deal with this
