### 2022-11-28 8524148
add some custom treatments for 2d delta_times and cases when there is a gt and spot
