### 2022-11-29 941f69a
add atl11 path parsing and docs
