### 2022-12-01 be0527a
update oa viz for polygons after spatial module change
