### 2022-12-02 9cbfd57
handle merging given non-unique ref_pt coordinates
