### 2022-12-14 #4912 @dimejimudele
Just a question for @kthyng, I could not find paper.md in the joss branch of the project repository, but I have performed my review responsibilities so far using the pdf [linked here](https://github.com/openjournals/joss-papers/blob/joss.04912/joss.04912/10.21105.joss.04912.pdf).

Any help?
