### 2022-12-14 #4912 @kthyng
@dimejimudele The JOSS paper is in a branch called `joss`. Within that it is pretty buried though! I found it here: https://github.com/icesat2py/icepyx/blob/joss/doc/source/tracking/paper.md
