### 2022-12-15 #4912 @dimejimudele
I have completed my review of this publication and the connected software tool.

I should say it is impressive work that has taken a lot of collaboration and time to deliver up to its current state.

[icepyx](https://github.com/icesat2py/icepyx) is a Python library for loading and processing ICESat-2 datasets. ICESat-2 is part of NASA's Earth Observing System and it provides data measuring ice sheet elevation and sea ice thickness, as well as land topography, vegetation characteristics, and clouds.

This tool enables faster workflows in many scientific domains.

- The authors provided easy-to-follow examples
- The installation guide provided (both through PyPI and Conda) works as provided in the documentation.
- Sufficient documentation has already been provided for this library
- There is a sufficient use case for this library
- It has taken beyond substantial scholarly effort to get this project to its current state

A minor clarification the authors might need to make is about the difference between [icesat2py](https://ui.adsabs.harvard.edu/abs/2019AGUFM.C31C1526S/abstract) and [icepyx](https://github.com/icesat2py/icepyx). It appears to be the case that the former is a previous name for the latter.

I do strongly recommend that this submission be published in JOSS.
