+++ .github/workflows/traffic_action.yml
@@ -17,7 +17,7 @@ jobs:
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
     # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
       with:
         ref: "traffic"
     
@@ -35,7 +35,7 @@ jobs:
     
     # Commits files to repository
     - name: Commit changes
-      uses: EndBug/add-and-commit@v8
+      uses: EndBug/add-and-commit@v9
       with:
         author_name: Jessica Scheick
         message: "GitHub traffic auto-update"
