### 2022-12-20 87bbcdc
only pip install needed packages instead of all requirements
