+++ .github/workflows/traffic_action.yml
@@ -30,7 +30,7 @@ jobs:
     # Custom steps to move files to another dir and format plot
     - name: Move files
       run: |
-        pip install -r requirements.txt
+        pip install matplotlib pandas
         python ./doc/source/tracking/traffic/traffic_data_mgmt.py
     
     # Commits files to repository
