### 2023-01-03 c62bf0a
add icesat2py clarification note
