### 2023-01-03 3ac9932
update contact page - discontinue regular meetings
