### 2023-01-03 c9d8cfd
simplify visualization bounding box code
