### 2023-01-09 75b2ed4
reactivate atl10 and atl12 tests and update sizes
