+++ icepyx/tests/test_visualization.py
@@ -79,8 +79,8 @@ def test_gran_paras(filename, expect):
         ("ATL06", ["2019-6-15", "2019-7-1"], [-64.5, -66, -63.5, -65], 3235),
         ("ATL07", ["2019-7-1", "2019-8-1"], [-65, -66, -64.5, -65], 6720),
         ("ATL08", ["2019-6-15", "2019-7-1"], [-18, 63, -17, 64], 852),
-        # ("ATL10", ["2019-8-1", "2019-9-1"], [-64, -67, -60, -60], 7375),
-        # ("ATL12", ["2019-7-1", "2019-10-1"], [-65.5, -65.5, -64.5, -65], 95),
+        ("ATL10", ["2019-8-1", "2019-9-1"], [-64, -67, -60, -60], 7100),
+        ("ATL12", ["2019-7-1", "2019-10-1"], [-65.5, -65.5, -64.5, -65], 95),
         ("ATL13", ["2019-6-1", "2019-12-1"], [-75, -51, -74, -50], 25),
     ],
 )
@@ -99,11 +99,15 @@ def test_visualization_date_range(product, date_range, bbox, expect):
         ("ATL06", [-64.5, -66, -63.5, -65], ["03"], ["1306"], 3235),
         ("ATL07", [-65, -66, -64.5, -65], ["04"], ["0186"], 6720),
         ("ATL08", [-18, 63, -17, 64], ["03"], ["1320"], 852),
-        # ("ATL10", [-64, -67, -60, -60], ["04"], ["0681"], 6015),
-        # ("ATL12", [-65.5, -65.5, -64.5, -65], ["05"], ["0041"], 95),
+        ("ATL10", [-64, -67, -60, -60], ["04"], ["0681"], 5175),
+        ("ATL12", [-65.5, -65.5, -64.5, -65], ["05"], ["0041"], 95),
         ("ATL13", [-75, -51, -74, -50], ["05"], ["0293"], 25),
     ],
 )
+
+# NOTE: in both failed tests, the issue is ATL13. They worked previously...
+
+
 def test_visualization_orbits(product, bbox, cycles, tracks, expect):
 
     region_viz = Visualize(
