+++ doc/source/community/contact.rst
@@ -3,9 +3,9 @@
 Contact Us
 ==========
 
-* Need help installing, running, or using `icepyx`? Ask for help on `Discourse <https://discourse.pangeo.io/c/science/icesat-2/16>`_ or `GitHub Discussions https://github.com/icesat2py/icepyx/discussions`_.
+* Need help installing, running, or using `icepyx`? Ask for help on `Discourse <https://discourse.pangeo.io/c/science/icesat-2/16>`_ or `GitHub Discussions <https://github.com/icesat2py/icepyx/discussions>`_.
 * Found a bug? Post an issue on `GitHub <https://github.com/icesat2py/icepyx/issues>`_!
-* Want to request or contribute a feature? Share your idea on `GitHub Discussions https://github.com/icesat2py/icepyx/discussions`_.
+* Want to request or contribute a feature? Share your idea on `GitHub Discussions <https://github.com/icesat2py/icepyx/discussions>`_.
 * Have a question or want to know more? Join us for a virtual meeting (see below).
 * Want to get involved? Do one or more of the above, or reach out to one of the dev team members individually. We're excited to hear your thoughts and provide help!
 
@@ -17,7 +17,7 @@ Virtual Meetings
 Our team (developers, users, scientists, educators) consists primarily of volunteers.
 We meet on an as-needed basis via Zoom to provide support, troubleshoot issues, and plan development and collaboration efforts.
 
-Our meetings are open to all, with upcoming meeting information available via `Discourse <https://discourse.pangeo.io/c/science/icesat-2/16>`_, `GitHub Discussions https://github.com/icesat2py/icepyx/discussions`_, or by request.
+Our meetings are open to all, with upcoming meeting information available via `Discourse <https://discourse.pangeo.io/c/science/icesat-2/16>`_, `GitHub Discussions <https://github.com/icesat2py/icepyx/discussions>`_, or by request.
 The :ref:`QUEST (Query Unify Explore SpatioTemporal) <api_quest_ref>`  team meets weekly on Mondays to co-work on integrating additional sensors into the icepyx workflow. 
 Please contact us if you are interested in joining the QUEST team.
 
