### 2023-01-17 0af53d6
assert bounding box type in Argo init, begin framework for unit tests
