+++ icepyx/tests/test_quest_argo.py
@@ -1,0 +1,14 @@
+import icepyx as ipx
+import pytest
+import warnings
+
+
+def test_available_profiles():
+	pass
+
+def test_no_available_profiles():
+	pass
+
+def test_valid_spatialextent():
+	pass
+
