+++ icepyx/tests/test_quest_BGCargo.py
@@ -1,0 +1,16 @@
+import icepyx as ipx
+import pytest
+import warnings
+
+
+def test_available_profiles():
+	pass
+
+def test_no_available_profiles():
+	pass
+
+def test_valid_BGCparams():
+	pass
+
+def test_invalid_BGCparams():
+	pass
