### 2023-01-17 f55fd61
assert bounding box type in Argo init, begin framework for unit tests
