+++ icepyx/core/read.py
@@ -437,11 +437,13 @@ class Read:
                 except NameError:
                     import random
 
-                    is2ds["gran_idx"] = [random.randint(900000, 999999)]
+                    is2ds["gran_idx"] = [random.randint(800000, 899998)]
                     warnings.warn("Your granule index is made up of random values.")
                     # You must include the orbit/cycle_number and orbit/rgt variables to generate
             except KeyError:
-                pass
+                # Added this when dealing with ATL11 - need to see if it breaks with other datasets
+                is2ds["gran_idx"] = [np.nanmax(is2ds["gran_idx"]) - 1]
+                # pass
 
             if hasattr(is2ds, "data_start_utc"):
                 is2ds = _make_np_datetime(is2ds, "data_start_utc")
@@ -456,14 +458,6 @@ class Read:
             else:
                 spot = track_str
 
-            # import re
-            # gt_str = re.match(r"gt[1-3]['r','l']", grp_path).group()
-            # spot = is2ref.gt2spot(gt_str, is2ds.sc_orient.values[0])
-            # # add a test for the new function (called here)!
-
-            # NEXT STEP: change out gt_str and spot in below code
-            # figure out gt_str --> track_str, spot_dim_name --> spot string, need spot value still
-
             grp_spec_vars = [
                 k
                 for k, v in wanted_dict.items()
@@ -471,12 +465,37 @@ class Read:
             ]
 
             # NEXT TODO: handle the case where it's the second time through a 2d delta_time...
+
+            # handle delta_times with 1 or more dimensions
+            idx_range = range(0, len(ds.delta_time.data))
+            # if hasattr(is2ds, "photon_idx"):
+
+            #     # if is2ds already has a 2d photon idx/delta time AND the current delta time does too
+            #     if np.ndim(ds.delta_time.data) > 1: # and np.ndim(is2ds.photon_idx) > 1:
+            #         # repeat the range the number of times needed, then transpose to match the shape of the existing photon_idx
+            #         photon_ids = (
+            #         np.broadcast_to([*idx_range], (np.shape(ds.delta_time)[1], np.shape(ds.delta_time)[0])).transpose()
+            #         + np.full_like(
+            #             ds.delta_time, np.max(is2ds.photon_idx), dtype="int64"
+            #         )
+            #         + 1
+            #     )
+            #     # the original case, where delta_time is 2d but the existing photon_idx is 1d
+            #     else:
+            #         photon_ids = (
+            #             range(0, len(ds.delta_time.data))
+            #             + np.full_like(
+            #                 ds.delta_time, np.max(is2ds.photon_idx), dtype="int64"
+            #             )
+            #             + 1
+            #         )
+            # else:
+            #     photon_ids = range(0, len(ds.delta_time.data))
+
             try:
                 photon_ids = (
-                    range(0, len(ds.delta_time.data))
-                    + np.full_like(
-                        ds.delta_time, np.max(is2ds.photon_idx), dtype="int64"
-                    )
+                    idx_range
+                    + np.full_like(idx_range, np.max(is2ds.photon_idx), dtype="int64")
                     + 1
                 )
             except AttributeError:
@@ -494,11 +513,9 @@ class Read:
                 )
                 .assign(gt=(("gran_idx", spot_dim_name), [[track_str]]))
                 .rename_dims({"delta_time": "photon_idx"})
-                # .set_index("delta_time")
                 .rename({"delta_time": "photon_idx"})
+                # .set_index("photon_idx")
             )
-
-            # try turning ref_pt into a non-dimension (and thus non-index?) coordinate???
 
             # handle cases where the delta time is 2d due to multiple cycles in that group
[…]
