### 2023-01-27 f9af8c9
update test failure notes
