### 2023-02-01 6ca5841
finish coding use of earthaccess (old code still commented)
