### 2023-02-02 8a725d0
Update README.rst
Co-authored-by: Zach Fair <48361714+zachghiaccio@users.noreply.github.com>
