### 2023-02-02 47c4119
Update doc/source/getting_started/origin_purpose.rst
