### 2023-02-06 27ab9d7
need to confirm spatial extent is bbox
