### 2023-02-06 83e0e94
begin test case for available profiles
