### 2023-02-06 d96c485
add tests for argo.py
