+++ icepyx/quest/dataset_scripts/argo.py
@@ -1,128 +1,136 @@
 from icepyx.quest.dataset_scripts.dataset import DataSet
-from icepyx.core.geospatial import geodataframe
+from icepyx.core.spatial import geodataframe
 import requests
 import pandas as pd
 import os
 import numpy as np
 
-class Argo(DataSet):
-	"""
-	Initialises an Argo Dataset object
-	Used to query physical Argo profiles
-	-> biogeochemical Argo (BGC) not included
 
-	Examples
+class Argo(DataSet):
+    """
+        Initialises an Argo Dataset object
+        Used to query physical Argo profiles
+        -> biogeochemical Argo (BGC) not included
+
+        Examples
     --------
     # example with profiles available
     >>> reg_a = Argo([-154, 30,-143, 37], ['2022-04-12', '2022-04-26'])
-	>>> reg_a.search_data()
-	>>> print(reg_a.profiles[['pres', 'temp', 'lat', 'lon']].head())
-	   pres    temp     lat      lon
-	0   3.9  18.608  33.401 -153.913
-	1   5.7  18.598  33.401 -153.913
-	2   7.7  18.588  33.401 -153.913
-	3   9.7  18.462  33.401 -153.913
-	4  11.7  18.378  33.401 -153.913
+        >>> reg_a.search_data()
+        >>> print(reg_a.profiles[['pres', 'temp', 'lat', 'lon']].head())
+           pres    temp     lat      lon
+        0   3.9  18.608  33.401 -153.913
+        1   5.7  18.598  33.401 -153.913
+        2   7.7  18.588  33.401 -153.913
+        3   9.7  18.462  33.401 -153.913
+        4  11.7  18.378  33.401 -153.913
 
-	# example with no profiles
-	>>> reg_a = Argo([-55, 68, -48, 71], ['2019-02-20', '2019-02-28'])
-	>>> reg_a.search_data()
-	Warning: Query returned no profiles
-	Please try different search parameters
-
-
-	See Also
-	--------
-	DataSet
-	GenQuery
-	"""
-	def __init__(self, boundingbox, timeframe):
-		super().__init__(boundingbox, timeframe)
-		assert self.spatial._ext_type == 'bounding_box'
-		self.profiles = None
+        # example with no profiles
+        >>> reg_a = Argo([-55, 68, -48, 71], ['2019-02-20', '2019-02-28'])
+        >>> reg_a.search_data()
+        Warning: Query returned no profiles
+        Please try different search parameters
 
 
-	def search_data(self, presRange=None, printURL=False):
-		"""
-		query dataset given the spatio temporal criteria
-		and other params specific to the dataset
-		"""
+        See Also
+        --------
+        DataSet
+        GenQuery
+    """
 
-		# builds URL to be submitted
-		baseURL = 'https://argovis.colorado.edu/selection/profiles/'
-		payload = {'startDate': self._start.strftime('%Y-%m-%d'),
-				   'endDate': self._end.strftime('%Y-%m-%d'),
-				   'shape': [self._fmt_coordinates()]}
-		if presRange:
-			payload['presRange'] = presRange
+    def __init__(self, boundingbox, timeframe):
+        super().__init__(boundingbox, timeframe)
+        assert self._spatial._ext_type == "bounding_box"
+        self.profiles = None
 
-		# submit request
-		resp = requests.get(baseURL, params=payload)
+    def search_data(self, presRange=None, printURL=False) -> str:
+        """
+        query dataset given the spatio temporal criteria
+        and other params specific to the dataset
+        """
 
-		if printURL:
-			print(resp.url)
+        # builds URL to be submitted
+        baseURL = "https://argovis.colorado.edu/selection/profiles/"
[…]
