### 2023-02-06 4e5f533
add test for new track type function
