### 2023-02-06 9226737
clean up and apply some of review suggestions
