### 2023-02-08 ef7ec3d
add icesat2py clarification note (#395)
Co-authored-by: Zach Fair <48361714+zachghiaccio@users.noreply.github.com>
