### 2023-02-09 191851f
Apply suggestions from code review
change path to pair track in docs

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
