### 2023-02-09 4bb4cf9
specify typeerror in try/except
