### 2023-02-09 8568fae
update failing test to match updated var name
