### 2023-02-13 4ec53cd
add typing, add example json, and use it to test parsing
