+++ icepyx/tests/test_quest_argo.py
@@ -1,4 +1,5 @@
 import icepyx as ipx
+import json
 import pytest
 import warnings
 from icepyx.quest.dataset_scripts.argo import Argo
@@ -48,11 +49,21 @@ def test_fmt_coordinates():
 
 
 def test_parse_into_df():
-    reg_a = Argo([-154, 30, -143, 37], ["2022-04-12", "2022-04-26"])
+    reg_a = Argo([-154, 54, -151, 56], ["2023-01-20", "2023-01-29"])
     reg_a.search_data()
+    obs_df = reg_a.profiles
 
-    pass
+    exp = 0
+    with open("./icepyx/tests/argovis_test_data2.json") as file:
+        data = json.load(file)
+        for profile in data:
+            exp = exp + len(profile["measurements"])
+
+    assert exp == len(obs_df)
 
     # goal: check number of rows in df matches rows in json
     # approach: create json files with profiles and store them in test suite
     # then use those for the comparison
+    # update: for some reason the file downloaded from argovis and the one created here had different lengths
+    # by downloading the json from the url of the search here, they matched
+    # next steps: testing argo bgc?
