### 2023-02-14 c6f1d88
docs: update CONTRIBUTORS.rst [skip ci]
