+++ CONTRIBUTORS.rst
@@ -8,58 +8,61 @@ Thanks goes to these wonderful people (`emoji key <https://allcontributors.org/d
 .. raw:: html
 
     <!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
-  <!-- prettier-ignore-start -->
-  <!-- markdownlint-disable -->
-  <table>
+<!-- prettier-ignore-start -->
+<!-- markdownlint-disable -->
+    <table>
+  <tbody>
     <tr>
-      <td align="center"><a href="https://nsidc.org"><img src="https://avatars.githubusercontent.com/u/47193922?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Amy Steiker</b></sub></a><br /><a href="#example-asteiker" title="Examples">💡</a> <a href="https://github.com/icesat2py/icepyx/commits?author=asteiker" title="Documentation">📖</a> <a href="#mentoring-asteiker" title="Mentoring">🧑‍🏫</a></td>
-      <td align="center"><a href="https://github.com/annavalentine"><img src="https://avatars.githubusercontent.com/u/65192768?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Anna Valentine</b></sub></a><br /><a href="#example-annavalentine" title="Examples">💡</a></td>
-      <td align="center"><a href="http://psc.apl.uw.edu/people/investigators/anthony-arendt/"><img src="https://avatars.githubusercontent.com/u/4993098?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Anthony Arendt</b></sub></a><br /><a href="https://github.com/icesat2py/icepyx/issues?q=author%3Aaaarendt" title="Bug reports">🐛</a> <a href="#business-aaarendt" title="Business development">💼</a> <a href="#eventOrganizing-aaarendt" title="Event Organizing">📋</a> <a href="#fundingFinding-aaarendt" title="Funding Finding">🔍</a></td>
-      <td align="center"><a href="https://github.com/wallinb"><img src="https://avatars.githubusercontent.com/u/35819999?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Bruce Wallin</b></sub></a><br /><a href="https://github.com/icesat2py/icepyx/commits?author=wallinb" title="Code">💻</a> <a href="https://github.com/icesat2py/icepyx/commits?author=wallinb" title="Documentation">📖</a> <a href="#ideas-wallinb" title="Ideas, Planning, & Feedback">🤔</a> <a href="https://github.com/icesat2py/icepyx/pulls?q=is%3Apr+reviewed-by%3Awallinb" title="Reviewed Pull Requests">👀</a> <a href="#question-wallinb" title="Answering Questions">💬</a></td>
-      <td align="center"><a href="http://dshean.github.io"><img src="https://avatars.githubusercontent.com/u/1103530?v=4?s=100" width="100px;" alt=""/><br /><sub><b>David Shean</b></sub></a><br /><a href="https://github.com/icesat2py/icepyx/issues?q=author%3Adshean" title="Bug reports">🐛</a> <a href="#mentoring-dshean" title="Mentoring">🧑‍🏫</a></td>
-      <td align="center"><a href="https://www.linkedin.com/in/landungsetiawan/"><img src="https://avatars.githubusercontent.com/u/17802172?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Don Setiawan</b></sub></a><br /><a href="#infra-lsetiawan" title="Infrastructure (Hosting, Build-Tools, etc)">🚇</a> <a href="#platform-lsetiawan" title="Packaging/porting to new platform">📦</a></td>
[…]
