### 2023-02-17 775dba1
Update icepyx/core/read.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
