### 2023-02-17 2b7a348
update test for updated function
