### 2023-02-17 bf6e632
update test for updated function
