### 2023-02-17 3595e39
update uml action to not install requirements
