+++ .github/workflows/uml_action.yml
@@ -15,7 +15,6 @@ jobs:
           ref: ${{ github.event.pull_request.head.ref }}
       - name: set up environment
         run: |
-          pip install -r requirements.txt
           pip install icepyx pylint 
           sudo apt-get install graphviz
       - name: run pyreverse
