### 2023-02-17 298204f
update read-in module for ATL11 (#398)
* remove sc_orient from atl11 required vars list

* introduce new function to determine if data product uses ground tracks, paths, or profiles and generalize code accordingly

* add some custom treatments for 2d delta_times and cases when there is a gt and spot

* add atl11 path parsing and docs

* handle merging given non-unique ref_pt coordinates

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
