+++ icepyx/core/is2ref.py
@@ -264,6 +264,7 @@ def _default_varlists(product):
         return common_list
 
 
+# dev goal: check and test this function
 def gt2spot(gt, sc_orient):
 
     assert gt in [
