+++ .github/workflows/uml_action.yml
@@ -15,6 +15,8 @@ jobs:
           ref: ${{ github.event.pull_request.head.ref }}
       - name: set up environment
         run: |
+          pip install -U pip
+          pip install -r requirements
           pip install icepyx pylint 
           sudo apt-get install graphviz
       - name: run pyreverse
