### 2023-02-17 7435246
add manual trigger option to test
