+++ .github/workflows/uml_action.yml
@@ -2,6 +2,7 @@ name: Update UML diagrams
 on:
   pull_request_review:
     types: [submitted]
+  workflow_dispatch:
 
 jobs:
   diagrams:
