### 2023-02-17 c9ad415
Revert "add manual trigger option to test"
This reverts commit 74352468c84cdb15c1bfbbde07ab98c9b449da29.
