### 2023-02-22 8e6b353
add .txt to requirements
