### 2023-02-25 #4912 @VaasuDevanS
Thanks for the opportunity to review this tool and apologies for my delay.

In addition to [dimejimudele](https://github.com/dimejimudele)'s review, I must say I am quite impressed with its functionality and ease of use. The tool's installation instructions and interface is very intuitive and straightforward to use. Users can easily install the package and use it to access and process the ICESat-2 NASA datasets. Installation works as expected and there is enough documentation and detailed change log about previous versions.

I believe this tool excels in the following components.

- **Data Download** - This tool makes it easy to download from NASA NSIDC DAAC quickly and efficiently without having to manually download data from the website.

- **Data Processing** - Workflows can be automated using this tool which can significant time when compared with processing the data manually.

- **Visualization** - One other functionality that comes out of the box is to perform visualization of the datasets which helps users to better understand the data.

Overall, I see icepyx useful for wide range of workflows, this tool helps users save a lot of time. I also strongly recommend that this submission should be published in JOSS.
