### 2023-02-28 3a3a7a1
remove email capabilities from earthdata_login and move entirely to order_granules. other earthaccess code edits and cleanup
