+++ icepyx/core/query.py
@@ -857,34 +857,39 @@ class Query(GenQuery):
 
     # test cases: uid, email supplied and not
 
-    def earthdata_login(self, uid=None, email=None, s3token=False, persist=False):
+    def earthdata_login(
+        self, uid=None, email=None, s3token=False, persist=False
+    ) -> None:
         """
-        Log in to NSIDC EarthData to access data. Generates the needed session and token for most
-        data searches and data ordering/download.
+        Authenticate with NASA EarthData to enable data ordering and download.
+
+        Authentication is completed using the [earthaccess library](https://nsidc.github.io/earthaccess/).
+        Generates the needed authentication sessions and tokens, including for cloud access.
 
         Parameters
         ----------
-        uid : string
+        uid : string, default None
             Earthdata login user ID
         email : string, default None
-            Email address. NSIDC will automatically send you emails about the status of your order.
+            Deprecated keyword for backwards compatibility.
         s3token : boolean, default False
             Generate AWS s3 ICESat-2 data access credentials
-
-        See Also
-        --------
-        Earthdata.Earthdata
+        persist : boolean, default False
+            Whether or not you would like your login credentials saved to a .netrc file.
 
         Examples
         --------
         >>> reg_a = ipx.Query('ATL06',[-55, 68, -48, 71],['2019-02-20','2019-02-28']) # doctest: +SKIP
-        >>> reg_a.earthdata_login(user_id,user_email) # doctest: +SKIP
+        >>> reg_a.earthdata_login(user_id) # doctest: +SKIP
         Earthdata Login password:  ········
         """
 
-        # next steps: clean this up now that the "try multiple options" is implemented in earthaccess
-        # then, migrate the testing suite for this to earthaccess
-        # will need to heavily update docs to point to the right spots of earthaccess
+        # next steps:
+        # write tests for this updated function (update current mock?)
+        # check email or not usage for order granules
+        # if appropriate, migrate the testing suite for this to earthaccess
+        # will need to heavily update docs and examples to point to the right spots of earthaccess
+        # remove the earthdata module
         # then can also try the s3 credential step...
 
         try:
@@ -892,23 +897,31 @@ class Query(GenQuery):
         except:
             pass
 
-        if os.environ.get("EDL_PASSWORD") != None:
-            pwd = os.environ.get("EDL_PASSWORD")
+        if os.environ.get("EARTHDATA_USERNAME") or os.environ.get("EDL_USERNAME"):
 
-            if os.environ.get("EARTHDATA_PASSWORD") != None:
+            if (
+                os.environ.get("EDL_PASSWORD") != None
+                and os.environ.get("EDL_USERNAME") != None
+            ):
+                pass
+
+            elif (
+                os.environ.get("EARTHDATA_PASSWORD") != None
+                and os.environ.get("EARTHDATA_USERNAME") != None
+            ):
+                uid = os.environ.get("EARTHDATA_USERNAME")
                 pwd = os.environ.get("EARTHDATA_PASSWORD")
                 warnings.warn(
-                    "Please update your environment variable names to 'EDL_PASSWORD'",
+                    "Please update your environment variable names to 'EDL_USERNAME' and 'EDL_PASSWORD'",
                     DeprecationWarning,
                 )
+                os.environ["EDL_PASSWORD"] = str(uid)
                 os.environ["EDL_PASSWORD"] = str(pwd)
 
             auth = earthaccess.login(strategy="environment")
 
         else:
-            auth = earthaccess.login(
-                strategy="interactive"
-            )  # , persist=persist) un-comment this once a new version of earthaccess is built
+            auth = earthaccess.login(strategy="interactive", persist=persist)
 
         # if s3token == False:
[…]
