+++ icepyx/tests/test_Earthdata.py
@@ -10,7 +10,6 @@ import pytest
 import re
 import shutil
 import warnings
-# from icepyx.core.Earthdata import Earthdata
 
 # PURPOSE: test different authentication methods
 @pytest.fixture(scope="module", autouse=True)
