### 2023-03-02 d263656
start at updating mocks and testing for earthdata_login function
