+++ icepyx/tests/conftest.py
@@ -8,8 +8,10 @@ def mock_settings_env_vars():
     with mock.patch.dict(
         "os.environ",
         {
-            "EARTHDATA_USERNAME": "icepyx_devteam",
-            "EARTHDATA_PASSWORD": "fake_earthdata_password",
+            "EDL_USERNAME": "icepyx_devteam",
+            "EDL_PASSWORD": "fake_earthdata_password",
+            "EARTHDATA_USERNAME": "icepyx_devteam2",
+            "EARTHDATA_PASSWORD": "fake_earthdata_password2",
             "EARTHDATA_EMAIL": "icepyx.dev@gmail.com",
         },
     ):
@@ -18,11 +20,21 @@ def mock_settings_env_vars():
 
 @pytest.fixture(scope="session")
 def username():
-    return os.environ.get("EARTHDATA_USERNAME")
+    return os.environ.get("EDL_USERNAME")
 
 
 @pytest.fixture(scope="session")
 def password():
+    return os.environ.get("EDL_PASSWORD")
+
+
+@pytest.fixture(scope="session")
+def dep_username():
+    return os.environ.get("EARTHDATA_USERNAME")
+
+
+@pytest.fixture(scope="session")
+def dep_password():
     return os.environ.get("EARTHDATA_PASSWORD")
 
 
