+++ icepyx/tests/test_Earthdata.py
@@ -1,13 +1,10 @@
 #!/usr/bin/env python
 """
-test_Earthdata.py (12/2021)
+test icepyx.core.query.Query.earthdata_login function
 """
-import getpass
 import netrc
 import os
-import posixpath
 import pytest
-import re
 import shutil
 import warnings
 
@@ -27,47 +24,47 @@ def setup_earthdata():
         os.remove(netrc_file)
 
 
-def capability_url(product, version):
-    return posixpath.join(
-        "https://n5eil02u.ecs.nsidc.org",
-        "egi",
-        "capabilities",
-        f"{product}.{version}.xml",
-    )
-
-
-def test_environment(username, password, email):
-    url = capability_url("ATL06", "005")
+@pytest.mark.test_dep
+def test_dep_environment(dep_username, dep_password):
     netrc_file = os.path.join(os.path.expanduser("~"), ".netrc")
     assert not os.access(netrc_file, os.F_OK)
-    assert Earthdata(username, email, url, password).login(attempts=1)
+    assert earthdata_login(dep_username, dep_password)
 
 
-def test_netrc(username, password, email):
+def test_environment(username, password):
+    netrc_file = os.path.join(os.path.expanduser("~"), ".netrc")
+    assert not os.access(netrc_file, os.F_OK)
+    assert earthdata_login(username, password)
+
+
+def test_netrc(username, password):
     # append to netrc file and set permissions level
     args = (username, "urs.earthdata.nasa.gov", password)
     netrc_file = os.path.join(os.path.expanduser("~"), ".netrc")
     with open(netrc_file, "a+") as f:
         f.write("machine {1} login {0} password {2}\n".format(*args))
         os.chmod(netrc_file, 0o600)
-    url = capability_url("ATL06", "005")
-    assert Earthdata(username, email, url, password).login(attempts=0)
+    assert earthdata_login(username, password)
 
 
-class Earthdata:
+def earthdata_login(
+    uid=None, pwd=None, email=None, s3token=False, persist=False
+) -> bool:
     """
-    Mocks the icepyx Earthdata login
+    Mocks the icepyx.core.query.Query.earthdata_login function
 
     Parameters
-    ----------
-    uid : str
-        Earthdata Login user name (user ID).
-    email : str
-        Complete email address, provided as a string.
-    password : str
-        Password for Earthdata registration associated with the uid.
-    capability_url : str
-        URL required to access Earthdata
+        ----------
+        uid : string, default None
+            Earthdata login user ID
+        pwd : string, default None
+            Earthdata login password
+        email : string, default None
+            Deprecated keyword for backwards compatibility.
+        s3token : boolean, default False
+            Generate AWS s3 ICESat-2 data access credentials
+        persist : boolean, default False
+            Whether or not you would like your login credentials saved to a .netrc file.
 
     Returns
     -------
@@ -75,61 +72,38 @@ class Earthdata:
         True if completed a successful "login", else False.
     """
 
-    def __init__(
-        self,
-        uid,
-        email,
-        capability_url,
-        pswd=os.environ.get("EARTHDATA_PASSWORD"),
-    ):
+    try:
+        url = "urs.earthdata.nasa.gov"
+        mock_uid, _, mock_pwd = netrc.netrc(netrc).authenticators(url)
+    except:
+        pass
 
-        assert isinstance(uid, str), "Enter your login user id as a string"
-        assert re.match(
-            r"[^@]+@[^@]+\.[^@]+", email
-        ), "Enter a properly formatted email address"
+    if os.environ.get("EARTHDATA_USERNAME") or os.environ.get("EDL_USERNAME"):
 
-        self.netrc = None
-        self.uid = uid
[…]
