### 2023-03-02 3c1fc60
continue trying to mock earthdata_login tests
