### 2023-03-03 #4912 @JessicaS11
Thanks for the review, @VaasuDevanS.

> I couldn't find obvious statements in the paper how this tool is compared to other packages (if there are any)

You're absolutely right that this is not definitively addressed in the paper. We do have a detailed [ICESat-2 Resource Guide](https://icepyx.readthedocs.io/en/latest/community/resources.html#) as part of our documentation that outlines other software and educational materials for working with ICESat-2. We'd be happy to add a reference and link to it within the JOSS paper if you'd recommend it.
