### 2023-03-03 #4912 @kthyng
@JessicaS11 Could you include in the paper a summary of the resources you have in the docs since it is a requirement for the JOSS paper specifically, in case the links change in the future? Plus that state of the field will presumably change over time and we want this to be a snapshot for right now. A reference and link would additionally be fine at the end of the summary.
