+++ requirements.txt
@@ -11,5 +11,6 @@ intake-xarray
 matplotlib
 numpy
 requests
+s3fs
 shapely
 xarray
