### 2023-03-06 0c10e16
add notes on install dev requirements
