### 2023-03-06 451747a
add note that complete install must be done with pip
