+++ setup.py
@@ -10,6 +10,7 @@ EXTRAS_REQUIRE = {
     "viz": ["geoviews >= 1.9.0", "cartopy >= 0.18.0", "scipy"],
 }
 EXTRAS_REQUIRE["complete"] = sorted(set(sum(EXTRAS_REQUIRE.values(), [])))
+# install with `pip install "icepyx[complete]"` There is no way to use this functionality with conda.
 
 setuptools.setup(
     name="icepyx",
