### 2023-03-07 2ecdd27
is2 resource guide edits
