### 2023-03-07 0c44012
is2 resource guide edits
