+++ doc/source/tracking/icepyx_pubs.bib
@@ -174,3 +174,30 @@ month={Oct},
 pages={5158},
 note={Attribution to icepyx from 15 Oct 2022 Twitter post, https://twitter.com/rblourenco/status/1581320878511382528}
 }
+
+
+% #########  Related Software ################
+
+@software{jp_swinski_2023_7705009,
+  author       = {JP Swinski and
+                  Eric Lidwa and
+                  Tyler Sutterley and
+                  David Shean and
+                  Joseph H Kennedy and
+                  Scott Henderson},
+  title        = {ICESat2-SlideRule/sliderule: v2.1.0},
+  month        = mar,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v2.1.0},
+  doi          = {10.5281/zenodo.7705009},
+  url          = {https://doi.org/10.5281/zenodo.7705009}
+}
+
+@software{OA,
+  author       = {Khalsa, S.J.S. and Borsa, A. and Nandigam, V. and others},
+  title        = {{OpenAltimetry} - rapid analysis and visualization of Spaceborne altimeter data},
+  year         = 2020,
+  publisher    = {Earth Sci Inform},
+  doi          = {10.1007/s12145-020-00520-2},
+  url          = {https://openaltimetry.org/}
