+++ doc/source/tracking/icepyx_pubs.bib
@@ -201,3 +201,4 @@ note={Attribution to icepyx from 15 Oct 2022 Twitter post, https://twitter.com/r
   publisher    = {Earth Sci Inform},
   doi          = {10.1007/s12145-020-00520-2},
   url          = {https://openaltimetry.org/}
+}
