### 2023-03-07 51e6171
try with @misc instead of @software in bib file
