+++ doc/source/tracking/icepyx_pubs.bib
@@ -178,7 +178,7 @@ note={Attribution to icepyx from 15 Oct 2022 Twitter post, https://twitter.com/r
 
 % #########  Related Software ################
 
-@software{jp_swinski_2023_7705009,
+@misc{SR,
   author       = {JP Swinski and
                   Eric Lidwa and
                   Tyler Sutterley and
@@ -194,7 +194,7 @@ note={Attribution to icepyx from 15 Oct 2022 Twitter post, https://twitter.com/r
   url          = {https://doi.org/10.5281/zenodo.7705009}
 }
 
-@software{OA,
+@misc{OA,
   author       = {Khalsa, S.J.S. and Borsa, A. and Nandigam, V. and others},
   title        = {{OpenAltimetry} - rapid analysis and visualization of Spaceborne altimeter data},
   year         = 2020,
