### 2023-03-13 0f10c8c
cleaned up draft implementation of new earthdata_login fn
