### 2023-03-13 6d33bca
remove Earthdata module
