+++ icepyx/core/Earthdata.py
@@ -1,174 +1,0 @@
-import earthaccess
-import requests
-import getpass
-import socket
-import netrc
-import re
-import os
-import json
-
-
-# DevNote: currently this class is not tested
-class Earthdata:
-    """
-    Initiate an Earthdata session for interacting
-    with the NSIDC DAAC.
-
-    Parameters
-    ----------
-    uid : string
-        Earthdata Login user name (user ID).
-    email : string, default None
-        Complete email address, provided as a string.
-    password : string (encrypted)
-        Password for Earthdata registration associated with the uid.
-    capability_url : string
-        URL required to access Earthdata
-
-    Returns
-    -------
-    Earthdata session object after a successful login
-    """
-
-    # ToDo: make the email optional throughout all steps in query
-    def __init__(
-        self,
-        uid,
-        email,  # might be able to make this not required but can still accept it as an input from query...
-        capability_url,
-        pswd=os.environ.get("EARTHDATA_PASSWORD"),
-    ):
-
-        assert isinstance(uid, str), "Enter your login user id as a string"
-
-        if email:
-            assert re.match(
-                r"[^@]+@[^@]+\.[^@]+", email
-            ), "Enter a properly formatted email address"
-
-        self.netrc = None
-        self.uid = uid
-        self.email = email
-        self.capability_url = capability_url
-        self.pswd = pswd
-
-    def _start_session(self):
-        # Request CMR token using Earthdata credentials
-        token_api_url = "https://cmr.earthdata.nasa.gov/legacy-services/rest/tokens"
-        # try initially with machine hostname
-        # revert to using localhost if gaierror exception
-        try:
-            ip = socket.gethostbyname(socket.gethostname())
-        except:
-            ip = socket.gethostbyname("localhost")
-
-        data = {
-            "token": {
-                "username": self.uid,
-                "password": self.pswd,
-                "client_id": "NSIDC_client_id",
-                "user_ip_address": ip,
-            }
-        }
-
-        response = None
-        response = requests.post(
-            token_api_url,
-            json=data,
-            headers={"Accept": "application/json", "Client-Id": "icepyx"},
-        )
-
-        print("response content")
-        print(response.content)
-        # check for a valid login
-        try:
-            json.loads(response.content)["token"]
-        except KeyError:
-            try:
-                print(json.loads(response.content)["errors"])
-            except KeyError:
-                print(
-                    "There are no error messages, but an Earthdata login token was not successfully generated"
-                )
-
-        token = json.loads(response.content)["token"]["id"]
-
-        session = requests.session()
-        s = session.get(self.capability_url)
-        response = session.get(s.url, auth=(self.uid, self.pswd))
-
-        self.session = session
-
-    def login(self, attempts=5):
-        """
-        This function tries to log the user in to Earthdata with the
-        information provided. It prompts the user for their Earthdata password,
-        but will only store that information within the active session.
-        If the login fails, it will ask the user to re-enter their
-        username and password up to a set number of times to try and log in.
-
-        Alternatively, you can create a .netrc file in your $HOME directory
-        with the following line:
-
-        machine urs.earthdata.nasa.gov login <uid> password <password>
-
-        Where <uid> is your NASA Earthdata user ID and <password> is your password
-        Then change the permissions of that file to 600
-        This will allow you to have read and write access to the file
-        No other user can access the file
-
-            $ chmod 600 ~/.netrc
-
-        The function checks for this file to retrieve credentials, prior to
-        prompting for manual input.
-
-        Parameters
-        ----------
-        attempts : int
-            Number of allowed retry attempts.
-
-        Examples
-        --------
[…]
