### 2023-03-14 1467f14
update earthdata_login syntax in docstring tests
