+++ icepyx/core/granules.py
@@ -513,7 +513,7 @@ class Granules:
         # Note: need to test these checks still
         if session is None:
             raise ValueError(
-                "Don't forget to log in to Earthdata using is2_data.earthdata_login(uid, email)"
+                "Don't forget to log in to Earthdata using query.earthdata_login()"
             )
             # DevGoal: make this a more robust check for an active session
 
