### 2023-03-14 996cd87
update earthdata_login syntax in docstring testsand docs
