### 2023-03-15 9681ab1
generalize to pass any accepted earthaccess.login() kwargs
