### 2023-03-15 7706ded
update earthdata tests
