### 2023-03-15 76ca9ed
update environment setup for travis to use conda to build environment (and not pin to an old version of python earthaccess doesn't support)
