### 2023-03-15 21df1ff
finish updating data access notebook for transition to earthaccess
