### 2023-03-15 9a62187
revert to using pip in test env because conda resolver takes too long
