+++ .travis.yml
@@ -12,8 +12,9 @@ before_install:
   - conda info -a
 
 install:
-  - conda create --name icepyx-env --channel conda-forge --file requirements.txt --file requirements-dev.txt proj geos
+  - conda create --name icepyx-env --channel conda-forge proj geos earthaccess
   - source activate icepyx-env
+  - pip install -r requirements.txt -r requirements-dev.txt
   - pip install -e .[complete]
 
 stages:
