### 2023-03-15 193851a
pin travis env to python 3.10
