### 2023-03-15 7586526
reintroduce s3url credentialing to earthdata_login
