### 2023-03-15 922f559
add daac to s3 credentials call
