### 2023-03-15 aef0f54
currently working version of ex and query with s3 credentials from earthaccess
