### 2023-03-15 a6970a0
clean up notes
