+++ requirements.txt
@@ -1,6 +1,6 @@
 backoff
 datashader
-earthaccess
+earthaccess>=0.5.1
 fiona
 geopandas
 h5netcdf
