### 2023-03-16 62e284c
fix typo in doc/source/tracking/paper.md
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
