### 2023-03-17 aec796a
manual solution for getting ATL15 s3 urls via icepyx
