### 2023-03-17 #4912 @VaasuDevanS
Thanks @JessicaS11. I can see those changes here - https://github.com/icesat2py/icepyx/commit/2ecdd27b10d08884f1daa227b5b1fb3262c70fb3. @kthyng Is it possible to see the same like this one - https://github.com/openjournals/joss-papers/blob/joss.04912/joss.04912/10.21105.joss.04912.pdf
