### 2023-03-20 d26c9a1
transition icepyx to using earthaccess library for authentication (#410)
* add earthaccess (formerly earthdata) as dependency
* remove Earthdata module
* update earthdata_login function to use earthaccess 
* update data access and other example notebooks
* update earthdata  and NSIDC login tests
* simplify cloud access example auth handling using earthaccess
