+++ doc/source/user_guide/documentation/components.rst
@@ -11,14 +11,6 @@ APIformatting
    :undoc-members:
    :show-inheritance:
 
-Earthdata
----------
-
-.. automodule:: icepyx.core.Earthdata
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 spatial
 ----------
 
