+++ .travis.yml
@@ -12,9 +12,8 @@ before_install:
   - conda info -a
 
 install:
-  - conda create --name icepyx-env --channel conda-forge python=3.7 proj geos
+  - conda create --name icepyx-env --channel conda-forge python=3.10 proj geos earthaccess
   - source activate icepyx-env
-
   - pip install -r requirements.txt -r requirements-dev.txt
   - pip install -e .[complete]
 
