### 2023-03-20 eaefa43
docs: update CONTRIBUTORS.rst [skip ci]
