### 2023-03-20 8a7224e
Update CONTRIBUTORS.rst
