+++ CONTRIBUTORS.rst
@@ -68,4 +68,4 @@ Thanks goes to these wonderful people (`emoji key <https://allcontributors.org/d
 
     <!-- ALL-CONTRIBUTORS-LIST:END -->
 
-    This project follows the `all-contributors <https://github.com/all-contributors/all-contributors>`_ specification. Contributions of any kind welcome!
+This project follows the `all-contributors <https://github.com/all-contributors/all-contributors>`_ specification. Contributions of any kind welcome!
