+++ icepyx/tests/test_behind_NSIDC_API_login.py
@@ -28,9 +28,10 @@ def session(reg):
         f.write("machine {1} login {0} password {2}\n".format(*args))
         os.chmod(netrc_file, 0o600)
 
-    ed_obj = reg.earthdata_login()
-    yield ed_obj._session
-    ed_obj._session.close()
+    reg.earthdata_login()
+    ed_obj = reg._session
+    yield ed_obj
+    ed_obj.close()
 
 
 ########## is2ref module ##########
