### 2023-03-28 #4912 @JessicaS11
@kthyng I am working on these tasks and have a question.
> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

The software version we submitted for review in Oct was 0.6.3, and we're now at 0.7.0. Should I create an intermediate release that is 0.6.3 + the revised JOSS submission, or can I put the revised JOSS paper on top of our current version and release that?
