### 2023-03-28 ca5df45
add CITATION.cff file
