### 2023-03-28 d5fccdc
update citation docs with Zenodo doi and 'icepyx Developers' as author
