+++ CITATION.rst
@@ -6,33 +6,50 @@ Citing icepyx
 icepyx
 ------
 
-This community and software is developed with the goal of supporting science applications. Thus, our contributors (including those who have developed the packages used within icepyx) and maintainers justify their efforts and demonstrate the impact of their work through citations.
+This community and software is developed with the goal of supporting science applications.
+Thus, our contributors (including those who have developed the packages used within icepyx) 
+and maintainers justify their efforts and demonstrate the impact of their work through citations.
 
-If you have used icepyx in your work, please consider citing our library:
-    Scheick, J. *et al.*, (2019). icepyx: Python tools for obtaining and working with ICESat-2 data.
-    https://github.com/icesat2py/icepyx.
-    
+If you have used icepyx in your work, please consider citing our library.
+We encourage you to use a version-specific citation and DOI (available from `Zenodo <https://zenodo.org/record/7754482>`_)
+to improve reproducibility and let users know the state of the software at the time of your analysis.
+
+A non-versioned citation of icepyx:
+    icepyx Developers, (2023). icepyx: Python tools for obtaining and working with ICESat-2 data.
+    Zenodo. https://doi.org/10.5281/zenodo.7754482
+
+
 A bibtex version for users working in Latex::
 
-    @Misc{icepyx,
+  @Misc{icepyx,
   author =    {Scheick, Jessica and others},
   organization = {icesat2py},
   title =     {{icepyx: Python} tools for obtaining and working with {ICESat-2} data},
-  year =      {2019--},
-  url = "https://github.com/icesat2py/icepyx"
-    }
+  year =      {2023},
+  doi = "https://doi.org/10.5281/zenodo.7754482",
+  publisher = {Zenodo},
+  url = "https://github.com/icesat2py/icepyx",
+  }
     
+for more information on the "icepyx Developers", please see our `Attribution Guidelines <https://icepyx.readthedocs.io/en/latest/contributing/attribution_link.html>`_.
+
 
 icepyx Dependencies
---------------------------
-If you have used one of the included packages to extend your data analysis capabilities within icepyx, please consider additionally citing that work, because it represents an independent software contribution to the open-source community. `SciPy <https://www.scipy.org/index.html>`_ provides a `helpful resource <https://www.scipy.org/citing.html>`_ for citing packages within the SciPy ecosystem (including Matplotlib, NumPy, pandas, and SciPy). Links to citation information for other commonly used packages are below.
+-------------------
+If you have used one of the included packages to extend your data analysis capabilities within icepyx, 
+please consider additionally citing that work, because it represents an independent software contribution to the open-source community. 
+`SciPy <https://www.scipy.org/index.html>`_ provides a `helpful resource <https://www.scipy.org/citing.html>`_ for citing 
+packages within the SciPy ecosystem (including Matplotlib, NumPy, pandas, and SciPy). 
+Links to citation information for other commonly used packages are below.
 
 - `fiona <https://github.com/Toblerity/Fiona/blob/master/CITATION.txt>`_
 - `GeoPandas <https://github.com/geopandas/geopandas/issues/812>`_
 - `Pangeo <https://github.com/pangeo-data/pangeo/issues/651>`_
 - `shapely <https://github.com/Toblerity/Shapely/blob/master/CITATION.txt>`_
+- `xarray <https://github.com/pydata/xarray/blob/main/CITATION.cff>`_
 
 
 ICESat-2 Data
 -------------
-ICESat-2 data citation depends on the exact dataset used. Citation information for each data product can be found through the `NSIDC website <https://nsidc.org/data/icesat-2/data-sets>`_.
+ICESat-2 data citation depends on the exact dataset used.
+Citation information for each data product can be found through the `NSIDC website <https://nsidc.org/data/icesat-2/data-sets>`_.
