+++ ATTRIBUTION.rst
@@ -3,48 +3,75 @@
 Attribution Guidelines
 ======================
 
-We are extremely grateful to everyone who has contributed to the success of the icepyx community, whether through direct contributions to or feedback about icepyx or as developers or maintainers of complimentary resources that are included within the icepyx ecosystem. This document outlines our goals to give appropriate attribution to all contributors to icepyx in ways that are fair and diverse and supportive of professional goals. To do so, we define broadly *contributions* as:
+We are extremely grateful to everyone who has contributed to the success of the icepyx community, 
+whether through direct contributions to or feedback about icepyx or as developers or maintainers of complimentary resources that are included within the icepyx ecosystem. 
+This document outlines our goals to give appropriate attribution to all contributors to icepyx in ways that are fair and diverse and supportive of professional goals. 
+To do so, we define broadly *contributions* as:
 
     Efforts towards achieving icepyx's goals, including writing code, tests, or documentation,
     development of example workflows, development, significant contributions, or maintenance of
     a tailored package that broadens the functionality of icepyx, feedback and suggestions,
     community building, etc.
 
-We use the terms "contributors", "developers", and "authors" interchangeably. We will recognize contributions in the following ways.
+We use the terms "contributors", "developers", and "authors" interchangeably.
+We aim recognize contributions in the following ways.
+
+
 
 Contributors List
 -----------------
-Anyone who has contributed a pull request to icepyx is welcome to add themselves to the ``CONTRIBUTORS.rst`` file located in the top level directory; the file is packaged and distributed with icepyx. This process is optional, but is the easiest way for us to say "thank you" to everyone who has helped this project.
+This project follows the `all-contributors <https://github.com/all-contributors/all-contributors>`_ specification. 
+When you contribute to icepyx for the first time or in a new way, you or a maintainer should have the `All Contributors bot
+open a PR <https://allcontributors.org/docs/en/bot/usage>`_` to recognize your contribution with `@all-contributors please add @<username> for <contributions>`.
+This will add you to the ``CONTRIBUTORS.rst`` file located in the top level directory; 
+the file is packaged and distributed with icepyx. 
+
 
 
 Example Workflows
 -----------------
-Many of the example workflows included within icepyx were developed by individuals or small teams for educational or research purposes. We encourage example developers to provide proper recognition for these efforts both within the notebook itself and by adding contributors to the `Contributors List`_ for attribution as describered herein.
+Many of the example workflows included within icepyx were developed by individuals or small teams for educational or research purposes. 
+We encourage example developers to provide proper recognition for these efforts both within the notebook itself and 
+by adding contributors to the `Contributors List`_ for attribution as describered herein.
 
 
 Version Release on Zenodo
 -------------------------
-When new releases of icepyx are archived on Zenodo, anyone who has contributed to icepyx will be invited to be an author. The list of potential authors will be generated using the `Contributors List`. Thus, if you have contributed to icepyx and would like to be included as an author, you *must* add your full name, affiliation ("Unaffiliated" is acceptable), and ORCID (optional) to ``CONTRIBUTORS.rst``.
+Each new release of icepyx is `archived on Zenodo <https://zenodo.org/record/7754482>`_.
+We consider each of our contributors to be a co-author, and collectively represent the icepyx authors in citations as "icepyx Developers".
+As described above, a complete list of co-authors and their contribution types is available via the `Contributors List`_.
 
[…]
