### 2023-03-29 #4912 @JessicaS11
@kthyng I tried to check off my completion of one of the items in the checklist, and it seems to have somehow also checked off one of the items in your list and (I'm pretty sure) made one of the items in my part of the checklist disappear. I'm not entirely sure what happened, but wanted to make sure I mentioned it.
