### 2023-03-30 #4912 @JessicaS11
>are the changes in v0.7.0 due to the JOSS review, or just separate development?

Both. The changes requested in the JOSS review to e.g. the documentation are in the release. Changes to the JOSS publication itself are in a separate branch and have not yet been merged into a released version of the software. The other changes between versions include bug fixes and the adoption of a new external library for handling authentication. Here's a [full list](https://icepyx.readthedocs.io/en/latest/user_guide/changelog/v0.7.0.html) of changes.
