### 2023-03-31 #4912 @kthyng
@JessicaS11 Ok it sounds like you could use v0.6.4 as the version to associate with JOSS and just update the metadata to match your JOSS paper for that one, or you could make an intermediate release like you said that is v0.6.4 but specifically labeled for JOSS.
