+++ ATTRIBUTION.rst
@@ -1,31 +1,40 @@
 .. _attribution_ref_label:
 
-Attribution Guidelines
-======================
+Recognizing Contributions
+=========================
 
-We are extremely grateful to everyone who has contributed to the success of the icepyx community, 
-whether through direct contributions to or feedback about icepyx or as developers or maintainers of complimentary resources that are included within the icepyx ecosystem. 
+We are extremely grateful to everyone who has contributed to the success of the icepyx community and software.
 This document outlines our goals to give appropriate attribution to all contributors to icepyx in ways that are fair and diverse and supportive of professional goals. 
-To do so, we define broadly *contributions* as:
+We define *contributions* broadly as:
 
-    Efforts towards achieving icepyx's goals, including writing code, tests, or documentation,
-    development of example workflows, development, significant contributions, or maintenance of
-    a tailored package that broadens the functionality of icepyx, feedback and suggestions,
-    community building, etc.
+    Efforts towards achieving icepyx's goals, including (1) writing code, tests, or documentation,
+    (2) development of example workflows, (3) development, significant contributions, or maintenance of
+    a tailored package that broadens the functionality of icepyx, (4) feedback and suggestions,
+    (5) community building, (6) etc.
 
-We use the terms "contributors", "developers", and "authors" interchangeably.
-We aim recognize contributions in the following ways.
+We recognize contributions in the following ways.
 
+  **Note**: These policies are not set in stone and may be changed to
+  accommodate the growth of the project or the preferences of the community.
 
 
 Contributors List
 -----------------
 This project follows the `all-contributors <https://github.com/all-contributors/all-contributors>`_ specification. 
-When you contribute to icepyx for the first time or in a new way, you or a maintainer should have the `All Contributors bot
-open a PR <https://allcontributors.org/docs/en/bot/usage>`_` to recognize your contribution with `@all-contributors please add @<username> for <contributions>`.
-This will add you to the ``CONTRIBUTORS.rst`` file located in the top level directory; 
-the file is packaged and distributed with icepyx. 
+When you contribute to icepyx for the first time or in a new way, you or a maintainer can use the `All Contributors bot
+to open a PR <https://allcontributors.org/docs/en/bot/usage>`_` to recognize your contribution.
+Comment on an existing PR with `@all-contributors please add @<username> for <contributions>`.
+This will add you (or your new contribution type) to the ``CONTRIBUTORS.rst`` file located in the top level directory; 
+the file is packaged and distributed with icepyx, so each release has a record of contributors and their contribution types.
 
+
+Changelog
+---------
+
+Each release includes a changelog of updates.
+Everyone who has made a commit since the last release is listed, with new contributors indicated.
+This list is automatically generated using a Sphinx extension; where available, full names are used.
+If the user's full name is not available on GitHub, their GitHub handle is used.
 
 
 Example Workflows
@@ -38,28 +47,39 @@ by adding contributors to the `Contributors List`_ for attribution as describere
 Version Release on Zenodo
 -------------------------
 Each new release of icepyx is `archived on Zenodo <https://zenodo.org/record/7754482>`_.
-We consider each of our contributors to be a co-author, and collectively represent the icepyx authors in citations as "icepyx Developers".
-As described above, a complete list of co-authors and their contribution types is available via the `Contributors List`_.
+
+Following the collaborative approach of `The Turing Way <https://the-turing-way.netlify.app/community-handbook/acknowledgement/acknowledgement-members.html>_`,
+we aim to encourage community leadership and shared ownership of icepyx.
+To this end, beginning with version 0.7.1 we collectively represent the icepyx authors in citations 
+(including Zenodo releases) as "The icepyx Developers".
+
[…]
