### 2023-04-04 e63979c
clarify version numbers for new zenodo release authorship
