+++ doc/source/tracking/icepyx_pubs.bib
@@ -41,8 +41,7 @@ note = {Abstract and poster (presented by Wei Ji Leong). American Geophysical Un
 title = {Collaborative Computational Resource Development around {ICESat-2} Data: the {icepyx} Community and Library},
 journal = {Earth and Space Science Open Archive},
 pages = {9},
-year = {2022}, 
-% for AGU Fall Meeting 2021
+year = {2022},
 DOI = {10.1002/essoar.10511316.1},
 url = {https://doi.org/10.1002/essoar.10511316.1},
 comment = {abstract, poster},
@@ -155,4 +154,23 @@ url = {https://www.sciencedirect.com/science/article/pii/S0034425716305089},
 author = {Thorsten Markus and Tom Neumann and Anthony Martino and Waleed Abdalati and Kelly Brunt and Beata Csatho and Sinead Farrell and Helen Fricker and Alex Gardner and David Harding and Michael Jasinski and Ron Kwok and Lori Magruder and Dan Lubin and Scott Luthcke and James Morison and Ross Nelson and Amy Neuenschwander and Stephen Palm and Sorin Popescu and CK Shum and Bob E. Schutz and Benjamin Smith and Yuekui Yang and Jay Zwally},
 keywords = {ICESat-2, Land ice, Sea ice, Vegetation, Climate change, Satellite mission},
 abstract = {The Ice, Cloud, and land Elevation Satellite (ICESat) mission used laser altimetry measurements to determine changes in elevations of glaciers and ice sheets, as well as sea ice thickness distribution. These measurements have provided important information on the response of the cryopshere (Earth's frozen surfaces) to changes in atmosphere and ocean condition. ICESat operated from 2003 to 2009 and provided repeat altimetry measurements not only to the cryosphere scientific community but also to the ocean, terrestrial and atmospheric scientific communities. The conclusive assessment of significant ongoing rapid changes in the Earth's ice cover, in part supported by ICESat observations, has strengthened the need for sustained, high accuracy, repeat observations similar to what was provided by the ICESat mission. Following recommendations from the National Research Council for an ICESat follow-on mission, the ICESat-2 mission is now under development for planned launch in 2018. The primary scientific aims of the ICESat-2 mission are to continue measurements of sea ice freeboard and ice sheet elevation to determine their changes at scales from outlet glaciers to the entire ice sheet, and from 10s of meters to the entire polar oceans for sea ice freeboard. ICESat carried a single beam profiling laser altimeter that produced ~70m diameter footprints on the surface of the Earth at ~150m along-track intervals. In contrast, ICESat-2 will operate with three pairs of beams, each pair separated by about 3km cross-track with a pair spacing of 90m. Each of the beams will have a nominal 17m diameter footprint with an along-track sampling interval of 0.7m. The differences in the ICESat-2 measurement concept are a result of overcoming some limitations associated with the approach used in the ICESat mission. The beam pair configuration of ICESat-2 allows for the determination of local cross-track slope, a significant factor in measuring elevation change for the outlet glaciers surrounding the Greenland and Antarctica coasts. The multiple beam pairs also provide improved spatial coverage. The dense spatial sampling eliminates along-track measurement gaps, and the small footprint diameter is especially useful for sea surface height measurements in the often narrow leads needed for sea ice freeboard and ice thickness retrievals. The ICESat-2 instrumentation concept uses a low energy 532nm (green) laser in conjunction with single-photon sensitive detectors to measure range. Combining ICESat-2 data with altimetry data collected since the start of the ICESat mission in 2003, such as Operation IceBri
[…]
