### 2023-04-06 1c53bb2
joss publication final (#420)
the final paper.md for our joss publication

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Tyler Sutterley <tsutterl@uw.edu>
Co-authored-by: Romina Piunno <romina.piunno@gmail.com>
Co-authored-by: aaarendt <arendta@uw.edu>
Co-authored-by: Kelsey Bisson <48059682+kelseybisson@users.noreply.github.com>
Co-authored-by: Zach Fair <48361714+zachghiaccio@users.noreply.github.com>
Co-authored-by: liuzheng-arctic <51420076+liuzheng-arctic@users.noreply.github.com>
Co-authored-by: Don Setiawan <landungs@uw.edu>
Co-authored-by: Tian Li <
[…]
