+++ doc/source/tracking/paper.md
@@ -1,0 +1,160 @@
+---
+title: 'icepyx: querying, obtaining, analyzing, and manipulating ICESat-2 datasets'
+tags:
+  - Python
+  - ICESat-2
+  - LiDAR
+  - elevation
+  - community
+  - cloud
+
+authors:
+  - name: Jessica Scheick^[Corresponding author] # note this makes a footnote
+    orcid: 0000-0002-3421-4459
+    affiliation: 1
+  - name: Wei Ji Leong
+    orcid: 0000-0003-2354-1988
+    affiliation: 2
+  - name: Kelsey Bisson
+    orcid: 0000-0003-4230-3467
+    affiliation: 3
+  - name: Anthony Arendt
+    orcid: 0000-0003-0429-6905
+    affiliation: 4
+  - name: Shashank Bhushan
+    orcid: 0000-0003-3712-996X
+    affiliation: 4
+  - name: Zachary Fair
+    orcid: 0000-0002-6047-1723
+    affiliation: 5
+  - name: Norland Raphael Hagen
+    orcid: 0000-0003-1994-1153
+    affiliation: 6
+  - name: Scott Henderson
+    orcid: 0000-0003-0624-4965
+    affiliation: 4
+  - name: Friedrich Knuth
+    orcid: 0000-0003-1645-1984
+    affiliation: 4
+  - name: Tian Li
+    orcid: 0000-0002-1577-4004
+    affiliation: 7
+  - name: Zheng Liu
+    orcid: 0000-0003-4132-8136
+    affiliation: 4
+  - name: Romina Piunno
+    orcid: 0009-0000-1144-0915
+    affiliation: 8
+  - name: Nitin Ravinder
+    affiliation: 9
+  - name: Landung "Don" Setiawan
+    orcid: 0000-0002-1624-2667
+    affiliation: 4
+  - name: Tyler Sutterley
+    orcid: 0000-0002-6964-1194
+    affiliation: 4
+  - name: JP Swinski
+    affiliation: 5
+  - name: Anubhav
+    orcid: 0000-0003-4017-2862
+    affiliation: 10
+
+# Note: first three authors are driving the publication. Additional contributors/authors are listed in alphabetical order by last name. Anyone who also contributes substantially to preparing the JOSS submission will be moved into ABC order after the first three and before the "non-publication" contributors to icepyx. Non-responsive coauthors will be removed from the list since their permission to be included was not granted.
+
+affiliations:
+ - name: University of New Hampshire, USA
+   index: 1
+ - name: Development Seed, USA
+   index: 2
+ - name: Oregon State University, USA
+   index: 3
+ - name: University of Washington, USA
+   index: 4
+ - name: NASA Goddard Space Flight Center, USA
+   index: 5
+ - name: CarbonPlan, USA
+   index: 6
+ - name: University of Bristol, UK
+   index: 7
+ - name: University of Toronto, Canada
+   index: 8
+ - name: University of Leeds, UK
+   index: 9
+ - name: University of Maryland, College Park, USA
+   index: 10
+date: 23 September 2022
+bibliography: icepyx_pubs.bib
+
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+aas-doi:
+aas-journal:
+---
+
+# Summary
+
+icepyx is both a software library and a community composed of ICESat-2 (NASA satellite) data users, developers, maintainers, and the scientific community.
+We are working together to develop a shared library of resources - including existing resources, new code, tutorials, and use-cases/examples - that simplify the process of querying, obtaining, analyzing, and manipulating ICESat-2 datasets to enable scientific discovery.
+
+# Statement of need
+
[…]
