+++ doc/source/tracking/icepyx_pubs.bib
@@ -1,50 +1,54 @@
 % #########  materials about icepyx ################
 @misc{2022_IS2-HW-tutorials,
-  author       = {Scheick, Jessica and
-                  Arendt, Anthony and
-                  Haley, Charley and
-                  Henderson, Scott and
-                  Koh, Jane and
-                  Setiawan, Don and
-                  Alterman, Naomi and
-                  Meyer, Joachim and
-                  Cristea, Nicoleta and
-                  Schweiger, Axel and
-                  Barciauskas, Aimee and
-                  Smith, Ben and
-                  Piunno, Romina and
-                  Shapero, Daniel and
-                  Fair, Zachary and
-                  Arndt, Philipp and
-                  Leong, Wei Ji and
-                  Sutterley, Tyler and
-                  Snow, Tasha and
-                  Beig, Mikala and
-                  Besso, Hannah and
-                  Liu, Zheng and
-                  Joughin, Ian and
-                  Bisson, Kelsey and
-                  Sauthoff, Wilson},
-  title        = {ICESat-2 Hackweek Website},
-  month        = apr,
-  year         = 2022,
-  note         = {If you use this book, please cite it as below.},
-  publisher    = {Zenodo},
-  version      = {2022.04.15},
-  doi          = {10.5281/zenodo.6462479},
-  url          = {https://doi.org/10.5281/zenodo.6462479}
+author       = {Scheick, J. and
+                Arendt, Anthony and
+                Haley, Charley and
+                Henderson, Scott and
+                Koh, Jane and
+                Setiawan, Don and
+                Alterman, Naomi and
+                Meyer, Joachim and
+                Cristea, Nicoleta and
+                Schweiger, Axel and
+                Barciauskas, Aimee and
+                Smith, Ben and
+                Piunno, Romina and
+                Shapero, Daniel and
+                Fair, Zachary and
+                Arndt, Philipp and
+                Leong, Wei Ji and
+                Sutterley, Tyler and
+                Snow, Tasha and
+                Beig, Mikala and
+                Besso, Hannah and
+                Liu, Zheng and
+                Joughin, Ian and
+                Bisson, Kelsey and
+                Sauthoff, Wilson},
+title        = {ICESat-2 Hackweek Website},
+month        = apr,
+year         = 2022,
+note         = {If you use this book, please cite it as below.},
+publisher    = {Zenodo},
+version      = {2022.04.15},
+doi          = {10.5281/zenodo.6462479},
+url          = {https://doi.org/10.5281/zenodo.6462479}
 }
 
-@misc{js2021agu,
-author = {Scheick, J and Bisson, K and Li, T and Leong, WJ and Arendt, A},
+@article{js2021agu,
+author = {Scheick, Jessica and Bisson, Kelsey and Li, Tian and Leong, Wei Ji and Arendt, Anthony},
 note = {Abstract and poster (presented by Wei Ji Leong). American Geophysical Union Fall Meeting, New Orleans, LA, USA. 13-17 December 2021.},
 title = {Collaborative Computational Resource Development around {ICESat-2} Data: the {icepyx} Community and Library},
-year = {2021},
+journal = {Earth and Space Science Open Archive},
+pages = {9},
+year = {2022},
+DOI = {10.1002/essoar.10511316.1},
+url = {https://doi.org/10.1002/essoar.10511316.1},
 comment = {abstract, poster},
 }
 
 @misc{js2020agu,
-author = {Scheick, J and Arendt, A and Heagy, L and Paolo, F and Perez, F and Steiker, A},
+author = {Scheick, J. and Arendt, A. and Heagy, L. and Paolo, F. and Perez, F. and Steiker, A.},
[…]
