### 2023-04-06 #4912 @JessicaS11
Hello @kthyng! I have completed all five of the tasks in the author checklist (I am unable to check them off).
Software release/tag (v0.6.4_JOSS): https://github.com/icesat2py/icepyx/tree/v0.6.4_JOSS
Zenodo link/doi: https://doi.org/10.5281/zenodo.7806097
Thanks for your hard work as lead editor!
