### 2023-04-11 #4912 @kthyng
List of paper comments:
* need a space between reference and parenthetical citation for several in the paper (like OpenAltimetry and its parenthetical reference)
* Lindsey Heagy is the correct spelling
* Make sure of proper capitalization in your references with {} around strings in your .bib file. For example, Li et al seems to not be properly capitalized.
