### 2023-04-11 #4912 @JessicaS11
Thanks @kthyng! Does JOSS have a recommended approach for applying these edits? I cannot update the files once they are on Zenodo (unless they agree to make an exception because I'm still within the one week window), and updating the repo won't port the changes over to Zenodo unless I create a new release (and thus change the release tag and and archive/DOI). Just want to make this as easy as possible for everyone!
