### 2023-04-11 #4912 @kthyng
I'd say just update the paper and don't do anything else — it doesn't impact the software so isn't critical for the version nor archive to be exact for the paper wording.
