+++ doc/source/tracking/icepyx_pubs.bib
@@ -25,7 +25,7 @@ author       = {Scheick, J. and
                 Joughin, Ian and
                 Bisson, Kelsey and
                 Sauthoff, Wilson},
-title        = {ICESat-2 Hackweek Website},
+title        = {{ICESat-2 Hackweek Website}},
 month        = apr,
 year         = 2022,
 note         = {If you use this book, please cite it as below.},
@@ -91,7 +91,7 @@ comment	= {tutorials}
 @misc{js2019agu,
 author = {Scheick, J. and Arendt, A. and Heagy, L. and Perez, F.},
 note = {Abstract and poster. American Geophysical Union Fall Meeting, San Francisco, California, USA. 9-13 December 2019.},
-title = {{Introducing} icepyx, an open source {Python} library for obtaining and working with {ICESat-2} data},
+title = {Introducing {icepyx}, an open source {Python} library for obtaining and working with {ICESat-2} data},
 year = {2019},
 comment = {abstract,poster},
 doi = {10.1002/essoar.10501423.1},
@@ -102,7 +102,7 @@ doi = {10.1002/essoar.10501423.1},
 
 @article{Bisson:2021,
 author = {Bisson, K. M. and Cael, B. B.},
-title = {How Are Under Ice Phytoplankton Related to Sea Ice in the Southern Ocean?},
+title = {How Are Under Ice Phytoplankton Related to Sea Ice in the {Southern Ocean}?},
 journal = {Geophysical Research Letters},
 volume = {48},
 number = {21},
@@ -113,7 +113,7 @@ year = {2021}
 }
 
 @inProceedings{Fernando:2021,
-  title = {Mapping the {{Diversity}} of {{Agricultural Systems}} in the {{Cuellaje Sector}}, {{Cotacachi}}, {{Ecuador Using ATL08}} for the {{ICESat-2 Mission}} and {{Machine Learning Techniques}}},
+  title = {Mapping the Diversity of Agricultural Systems in the {{Cuellaje Sector, Cotacachi, Ecuador}} Using {{ATL08}} for the {{ICESat-2 Mission}} and Machine Learning Techniques},
   booktitle = {Computational {{Science}} and {{Its Applications}} \textendash{} {{ICCSA}} 2021},
   author = {Fernando, Garrido},
   editor = {Gervasi, Osvaldo and Murgante, Beniamino and Misra, Sanjay and Garau, Chiara and Ble{\v c}i{\'c}, Ivan and Taniar, David and Apduhan, Bernady O. and Rocha, Ana Maria A. C. and Tarantino, Eufemia and Torre, Carmelo Maria},
@@ -129,7 +129,7 @@ year = {2021}
 
 @Article{Li:2020,
 AUTHOR = {Li, T. and Dawson, G. J. and Chuter, S. J. and Bamber, J. L.},
-TITLE = {Mapping the grounding zone of Larsen~C Ice Shelf, Antarctica, from ICESat-2 laser altimetry},
+TITLE = {Mapping the grounding zone of {Larsen~C Ice Shelf, Antarctica}, from {ICESat-2} laser altimetry},
 JOURNAL = {The Cryosphere},
 VOLUME = {14},
 YEAR = {2020},
@@ -143,7 +143,7 @@ DOI = {10.5194/tc-14-3629-2020}
 % #########  Articles about ICESat-2 ################
 
 @article{is2,
-title = {The Ice, Cloud, and land Elevation Satellite-2 (ICESat-2): Science requirements, concept, and implementation},
+title = {The {Ice, Cloud, and land Elevation Satellite-2 (ICESat-2)}: Science requirements, concept, and implementation},
 journal = {Remote Sensing of Environment},
 volume = {190},
[…]
