### 2023-04-11 6198e73
Update doc/CITATION.cff
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
