### 2023-04-12 62ff7e9
parse citations list into categories and add missing ones
